### 2024-11-09 #7444 @editorialbot
[PRE REVIEW]: FDApy: a Python package for functional data
**Submitting author:** @StevenGolovkine (<a href="http://orcid.org/0000-0002-5994-2671">Steven Golovkine</a>)
**Repository:** https://github.com/StevenGolovkine/FDApy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
