### 2024-11-09 #7444 @lrnv
@crvernon Hey thanks for the proposal. I'd rather not since I'm not a python guy and I don't know much about functional data :/
