### 2024-11-25 #7526 @mstimberg
at the top of a new comment in this thread.

There are additional guidelines and links in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/7526#issue-2689933928). Note that we usually ask reviewers to complete a first review within 4–6 weeks, but of course we know that this might not be feasible when end-of-year holidays get into the way. I will have limited availability from Dec 23–Jan 1st, myself.

Please feel free to ping me (@mstimberg) if you have any questions/concerns. Looking forward to working with you all!
