### 2024-11-25 #7526 @mstimberg
@StevenGolovkine no rush, but please have a look at the editorialbot comments about DOIs [above](https://github.com/openjournals/joss-reviews/issues/7526#issuecomment-2497395518) and add them to your bibliography list where appropriate.
