+++ joss/paper.bib
@@ -3,7 +3,6 @@
   title = {{{PEP8}} - {{Style}} Guide for {{Python}} Code},
   author = {{van Rossum}, Guido and Warsaw, Barry and Coghlan, Alyssa},
   year = {2001},
-  number = {8},
 }
 
 @book{ramsayFunctionalDataAnalysis2005,
@@ -31,7 +30,6 @@
   keywords = {functional-data-analysis,functional-time-series,hilbert-space-theory},
 }
 
-
 @book{kokoszkaIntroductionFunctionalData2017,
   title = {Introduction to {{Functional Data Analysis}}},
   author = {Kokoszka, Piotr and Reimherr, Matthew},
@@ -42,21 +40,6 @@
   keywords = {functional-data-analysis,functional-regression,functional-time-series,hilbert-space-theory},
 }
 
-@article{warmenhovenBivariateFunctionalPrincipal2019,
-  title = {Bivariate Functional Principal Components Analysis: Considerations for Use with Multivariate Movement Signatures in Sports Biomechanics},
-  shorttitle = {Bivariate Functional Principal Components Analysis},
-  author = {Warmenhoven, John and Cobley, Stephen and Draper, Conny and Harrison, Andrew and Bargary, Norma and Smith, Richard},
-  year = {2019},
-  month = jan,
-  journal = {Sports Biomechanics},
-  volume = {18},
-  number = {1},
-  pages = {10--27},
-  publisher = {{Routledge}},
-  urldate = {2023-09-15},
-  keywords = {biomechanics,functional-data-analysis,rowing,sport-science},
-}
-
 @article{songSparseMultivariateFunctional2022,
   title = {Sparse Multivariate Functional Principal Component Analysis},
   author = {Song, Jun and Kim, Kyongwon},
@@ -65,16 +48,6 @@
   volume = {11},
   number = {1},
   pages = {e435},
-}
-
-@article{golovkineClusteringMultivariateFunctional2022,
-  title = {Clustering Multivariate Functional Data Using Unsupervised Binary Trees},
-  author = {Golovkine, Steven and Klutchnikoff, Nicolas and Patilea, Valentin},
-  year = {2022},
-  month = apr,
-  journal = {Computational Statistics \& Data Analysis},
-  volume = {168},
-  pages = {107376},
 }
 
 @article{wangFaultDetectionBatch2015,
@@ -97,6 +70,7 @@
   number = {522},
   pages = {649--659},
   publisher = {{Taylor \& Francis}},
+  doi = {10.1080/01621459.2016.1273115}
 }
 
 @misc{golovkineUseGramMatrix2023,
@@ -106,9 +80,8 @@
   month = jun,
   number = {arXiv:2306.12949},
   primaryclass = {stat},
-  publisher = {{arXiv}},
+  publisher = {{arXiv}}
 }
-
 
 @inproceedings{allenMultiwayFunctionalPrincipal2013,
   title = {Multi-Way Functional Principal Components Analysis},
@@ -117,6 +90,7 @@
   year = {2013},
   month = dec,
   pages = {220--223},
+  doi = {10.1109/camsap.2013.6714047}
 }
 
 @misc{ramos-carrenoScikitfdaPythonPackage2023,
@@ -196,7 +170,7 @@
[…]
