+++ joss/paper.bib
@@ -15,6 +15,7 @@
   urldate = {2023-07-07},
   langid = {english},
   keywords = {functional-data-analysis},
+  doi = {10.1007/b98888}
 }
 
 @book{horvathInferenceFunctionalData2012a,
@@ -38,16 +39,6 @@
   publisher = {{Chapman and Hall/CRC}},
   address = {{New York}},
   keywords = {functional-data-analysis,functional-regression,functional-time-series,hilbert-space-theory},
-}
-
-@article{songSparseMultivariateFunctional2022,
-  title = {Sparse Multivariate Functional Principal Component Analysis},
-  author = {Song, Jun and Kim, Kyongwon},
-  year = {2022},
-  journal = {Stat},
-  volume = {11},
-  number = {1},
-  pages = {e435},
 }
 
 @article{wangFaultDetectionBatch2015,
@@ -78,9 +69,9 @@
   author = {Golovkine, Steven and Gunning, Edward and Simpkin, Andrew J. and Bargary, Norma},
   year = {2023},
   month = jun,
-  number = {arXiv:2306.12949},
   primaryclass = {stat},
-  publisher = {{arXiv}}
+  publisher = {{arXiv}},
+  doi = {10.48550/arXiv.2306.12949}
 }
 
 @inproceedings{allenMultiwayFunctionalPrincipal2013,
@@ -127,7 +118,8 @@
   number = {118},
   pages = {1--6},
   issn = {1533-7928},
-  urldate = {2024-01-10}
+  urldate = {2024-01-10},
+  doi = {10.5555/3455716.3455834}
 }
 
 @misc{ramsayFdaFunctionalData2023,
@@ -321,18 +313,6 @@
   number = {8},
 }
 
-@book{ramsayFunctionalDataAnalysis2005,
-  title = {Functional {{Data Analysis}}},
-  author = {Ramsay, J. O. and Silverman, B. W.},
-  year = {2005},
-  series = {Springer {{Series}} in {{Statistics}}},
-  publisher = {{Springer}},
-  address = {{New York, NY}},
-  urldate = {2023-07-07},
-  langid = {english},
-  keywords = {functional-data-analysis},
-}
-
 @book{horvathInferenceFunctionalData2012a,
   title = {Inference for {{Functional Data}} with {{Applications}}},
   author = {Horv{\'a}th, Lajos and Kokoszka, Piotr},
@@ -344,6 +324,7 @@
   urldate = {2023-09-19},
   langid = {english},
   keywords = {functional-data-analysis,functional-time-series,hilbert-space-theory},
+  doi = {10.1007/978-1-4614-3655-3}
 }
 
 
@@ -355,6 +336,7 @@
   publisher = {{Chapman and Hall/CRC}},
   address = {{New York}},
   keywords = {functional-data-analysis,functional-regression,functional-time-series,hilbert-space-theory},
+  doi = {10.1201/9781315117416}
 }
 
 @article{warmenhovenBivariateFunctionalPrincipal2019,
@@ -381,6 +363,7 @@
   volume = {11},
   number = {1},
   pages = {e435},
+  doi = {10.1002/sta4.435}
 }
 
 @article{golovkineClusteringMultivariateFunctional2022,
@@ -403,18 +386,6 @@
   volume = {149},
   pages = {78--89},
   doi = {10.1016/j.chemolab.2015.09.018}
-}
-
-@article{happMultivariateFunctionalPrincipal2018,
-  title = {Multivariate {{Functional Principal Component Analysis}} for {{Data Observed}} on {{Different}} ({{Dimensional}}) {{Domains}}},
-  author = {Happ, Clara and Greven, Sonja},
[…]
