+++ joss/paper.bib
@@ -195,7 +195,7 @@
   month = dec,
   urldate = {2024-01-10},
   copyright = {GPL-3},
-  doit = {10.32614/cran.package.fdasrvf}
+  doi = {10.32614/cran.package.fdasrvf}
 }
 
 @article{happ-kurzObjectOrientedSoftwareFunctional2020,
@@ -311,6 +311,7 @@
   author = {{van Rossum}, Guido and Warsaw, Barry and Coghlan, Alyssa},
   year = {2001},
   number = {8},
+  url = {https://peps.python.org/pep-0008/}
 }
 
 @book{horvathInferenceFunctionalData2012a,
