### 2024-12-08 StevenGolovkine/FDApy#19 @vnmabus
Issues with the paper
Hi, @StevenGolovkine. I want to first congratulate you for your work in this package and your submission to JOSS. As the author of scikit-fda, is for me a pleasure to review the work of a fellow developer in the field of functional data analysis in Python.

I create this issue to group all my observations with respect to the submitted paper, in relation with the open review in https://github.com/openjournals/joss-reviews/issues/7526. I will similarly create other issues for other relevant topics (documentation, functionality, performance, etc) as needed during the review process.

First, a list of typos:

"Most of the R packages that implements" -> "Most of the R packages that implement"
"FDApy supports the analysis diverse types of functional data" -> "FDApy supports the analysis of diverse types of functional data"
"We provide installation guides, algorithm introductions, examples of using the package" -> "We provide installation guides, algorithm introductions, and examples of using the package"
"It can be installed with pip install FDApy" -> "It can be installed with `pip install FDApy`" (use a monospace font for code)
"numpy" -> "NumPy"
"Steven Golovkine wish to thank" -> "Steven Golovkine wishes to thank"

The paper "Scikit-fda: A Python Package for Functional Data Analysis" is published by JSS (https://www.jstatsoft.org/article/view/v109i02). Please, use the updated reference:

```bibtex
@article{ramos-carreno++_2024_scikit-fda,
  author = {Ramos-Carreño, Carlos and Torrecilla, José L. and Carbajo Berrocal, Miguel and Marcos Manchón, Pablo and Suárez, Alberto},
  doi = {10.18637/jss.v109.i02},
  journal = {Journal of Statistical Software},
  month = may,
  number = {2},
  pages = {1--37},
  title = {{scikit-fda: A Python Package for Functional Data Analysis}},
  url = {https://www.jstatsoft.org/article/view/v109i02},
  volume = {109},
  year = {2024}
}
```

Now, more observations with respect to the content:

1. The description of the summary in the review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html) says "A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.". The summary in the paper talks only about the discipline of FDA and it does not mention the software itself. (To be honest, I am a bit conflicted about this given that the example paper at https://joss.readthedocs.io/en/latest/example_paper.html has the same problem, but I still think that at least a sentence summarizing the software would be good in this section).
2. For scikit-fda is mentioned that "they only accept multivariate functional data defined on the same domain". This is not completely true: for linear regression it is possible to use functions defined in different domains, or even mixed data (mixing functional and multivariate data). This is done by accepting a Pandas dataframe as input and integrating the functional datatypes with Pandas using the [`ExtensionArray`](https://pandas.pydata.org/docs/reference/api/pandas.api.extensions.ExtensionArray.html) functionality avaliable in Pandas, so that functional columns are accepted.
For the purposes of this paper, I would say it is enough to add "in most cases" to the sentence to make it correct (for now, we plan to extend this in the future).
3. I would state more clearly which functionality **is** available in FDApy. For example, you seem to have smoothing functionality, but this is not stated anywhere. Also, it is not clear the kinds of representations that are available for functional data (e.g. you have support for basis representation, but it is also not mentioned anywhere).
4. There seem to be other papers that use and cite this software, and could be referenced (it is up to you, but it makes clear that the software **is** useful for other researchers apart from your research group). A cursory search in Google Scholar shows that the following research articles use the software, and some others reference it:
    - [Detecting differences in gait initiation between older adult fallers and non-fallers through multivariate functional principal component analysis](https://www.sciencedirect.com/science/article/pii/S0021929022003839?casa_token=6Qb7_zNU8qMAAAAA:aoBMxCUxPuzSzgair7Er-n1Svh4dwhboULI339kT58dYLtcMQeyNr6v6rWKiMpUXtJPEIcldBiE)
    - [Learning domain-specific cameras (Thesis)](https://purl.stanford.edu/zm136ny2176)
