### 2024-12-08 StevenGolovkine/FDApy#20 @vnmabus
Issues with documentation in main page and README
A couple of issues with the main page of the documentation and the README (which are mostly the same), in relation with the open review in JOSS (https://github.com/openjournals/joss-reviews/issues/7526).

1. It is mentioned that "Functional Data Analysis, usually referred as FDA, concerns the field of Statistics that deals with discrete observations of continuous d-dimensional functions." However, I do not think that the fact that the observations are usually discrete has anything to do with FDA itself, being instead an artifact of the usual measurement apparatus. The literature of FDA often considers the functions as a whole, independently on how are they represented inside a computer. For example, if one could analogically transform the data to a basis expansion, such as Fourier, and input that into the computer, the observations would not be "discrete" anymore, but FDA could deal with those just fine.
2. "It includes methods for different dimensional data as well as irregularly sampled functional data." Which methods? Please explain the areas that are covered by the package, e.g. dimensionality reduction, representation and smoothing.
3. I would expand a bit the explanation given, so that a person that is new to the field of FDA can easily understand which functionalities are provided. Right now, the explanation is too short and references some concepts, such as "different dimensional data" or "irregularly sampled functional data" which may not be familiar for a newcomer. Note that FDA methods can be applied in a variety of scientific fields whose practitioners may not have an in-depth knowledge about the terminology used here.
4. If the software is designed to be compatible/interoperable with other software, such as scikit-learn, it would be good to state that here.
5. Not sure if it is ommited on purpose, giving that it is mostly an internal detail, but the package `lazy_loader` is missing from the "Requirements" section.
6. I would remove the autogenerated "Indices and tables" section: it is ugly and not very useful.
7. The Pydata Sphinx theme that you are using provides means to have links on the top bar of the docs. I suggest to at least provide a link to the Github repo.
