### 2024-12-08 StevenGolovkine/FDApy#21 @vnmabus
Issues with community guidelines
A couple of issues with the community guidelines, in relation with the open review in JOSS (https://github.com/openjournals/joss-reviews/issues/7526).

I see that you have a [CONTRIBUTING.rst](https://github.com/StevenGolovkine/FDApy/blob/master/CONTRIBUTING.rst) file. This is great to foster collaboration.

However, there are some ways in which this can be improved:

1. "credit will always be given" How is credit given? I do not see any THANKS.txt, [allcontributors](https://allcontributors.org/), list of contributors in the docs or changelog or similar.
2. Some explanations about how to set up your local copy for development are outdated: they reference the workflow using `setup.py` which does not longer exist. They should probably be fixed to use `pip install -e` and `pytest` respectively.
3. You can create a "Code of conduct" in order to state the behavior norms that need to be respected by contributors and users. Github can create one that meets basic requirements for you.
4. Some of the recommendations that are given for creating issues and PRs can be better enforced with the use of issue templates and pull request templates (if creating them from zero is too daunting you can freely adapt the ones from scikit-fda).
5. "The pull request should work for Python 3.9 and for PyPI." What does "for PyPI" mean in this context?
6. If you have a particular philosophy regarding Git history and the use of rebase, it would be good to state that here.
