### 2024-12-08 StevenGolovkine/FDApy#22 @vnmabus
Issues with API reference
A couple of issues with the API reference, in relation with the open review in JOSS (https://github.com/openjournals/joss-reviews/issues/7526).

General:
1. The API references part of the documentation is completely autogenerated, with all functionality for each module in one page. I think it would be way better that each class/function is documented in its own page, and referenced from the documentation of the module page. In that way, the documentation page of each module can have an overview explanation of the functionality present in the module and how does it relate with each other, as well as with the other parts of FDApy, while the internal details of each class/function do not clutter the discussion.
2. It would be great to have links from the documentation of a particular class/function to their implementation in the code.
3. It is possible for Sphinx to use the type hints for the documentation, in case you didn't know, so you do not have to repeat yourself in the docstring.
4. The order in which the attributes and methods are displayed in the documentation seems a bit arbitrary, with them mixing in some cases. I would try to put all attributes before all methods if possible.
5. For some functions, the arguments and/or return values are not documented, and they are not always at all obvious. I won't mention all the cases individually, just these in which the args are present and the return not or vice-versa, or cases in which I had particular difficulties.
6. I guess from the documentation that vector-valued functions are not supported, is that right?

Representation:
1. The name `n_dimension` seems a bit ambiguous when applied to functions, as it is not clear from the name if it is the input dimension, the output dimension, or the dimension of an "array" of functions.
1. In my experience, saying "observations" or "samples" with functional data is a bit ambiguous, as it is not always clear if one is referring to the observed values of the functions or to the functions themselves. I suggest using "functional observations" in the docstring for `n_obs` then, as it can serve to disambiguate the meaning.
2. The return value of `concatenate` seems not to be documented.
3. What is "long format"?
4. What is the "order" in the noise variance?
5. It is also not clear what `method_smoothing` has to do with `mean` and `center`. I guess it is used just for irregularly sampled data?
6. Functions may have different norms. I guess `norm` is the $L^2$ norm? Maybe it is better to say so explicitly, and even better to include the equation.
7. (Out of personal curiosity) Is standardization really used for functional data? Does it not alter the nature of the functional data itself?
8. From the docs is not clear at all what `rescale` is supposed to do.
9. What is `argvals_stand`?
10. Typo: "BasisFunctonalData" -> "BasisFunctionalData". Note that Sphinx also allows to reference other classes/functions and it creates an hyperlink.
11. "Be careful that it will not work if the curves are not sampled on [0, 1]". Then I am not sure this is a good default value. Why not forcing the user to specify a bandwith?
12. Typo: "Ramsey" -> "Ramsay" (several occurrences)
13. There is a "TODO" message in the docs.
14. "Be careful that we will not check if all the elements have the same type. It is possible to create MultivariateFunctionalData containing both Dense and Iregular functional data. However, only this two types are allowed to be in the list." -> Not true, given that data in a basis expansion is also allowed.
15. I would appreciate a description of the available basis, and when to use each one.

Preprocessing:
In general this module is better documented than the others, maybe because it contains the core functionality of the package.
1. Typo: "eigencoponents" -> "eigencomponents"

Visualization:
1. The documentation seems to be wrongly generated, as it references a non-stated plot function.
2. Is it not possible to plot functions represented as a basis expansion?

Simulation:
1. "The sampling points have to be regularly spaced. Otherwise, the covariance of the generated data will not be the good one." -> This should be improved, or at least checked for and raise an exception.

Miscellaneous:
1. The format of the CSV file should be specified, as well as the condition for returning dense vs sparse data.
