### 2024-12-08 StevenGolovkine/FDApy#23 @vnmabus
Issues with type hinting
A couple of issues with the type hinting, in relation with the open review in JOSS (https://github.com/openjournals/joss-reviews/issues/7526).

1. Given the presence of `Type[FunctionalData]` when the correct type would have been `FunctionalData`, I guess you are not using a type-checker such as MyPy/Pyright to validate your type hints. I encourage you to do so, as type-hinting without validation can create more confusion than anything else. It can also help you to detect bugs earlier, even before running tests.
2. If you use `from __future__ import annotations`, the annotations are not evaluated and you can use in Python 3.9 the pipe ("|") syntax from PEP 604 (available normally on versions above Python 3.10) to get rid of `Optional` and `Union`. Alternatively you can also drop support for Python 3.9, which is recommended anyway if you follow [SPEC 0](https://scientific-python.org/specs/spec-0000/).
3. For functions that accept a small set of valid strings, you can use the `Literal` type in the `typing` module, e.g. `method: Literal["PS", "LP"]`.
4. For functions that require that some arguments and/or return value have exactly the same type you can use the `TypeVar` class in the `typing` module.
