### 2024-12-20 StevenGolovkine/FDApy#25 @vnmabus
Issues with examples
A couple of issues with the examples, in relation with the open review in JOSS (openjournals/joss-reviews#7526).

1. The examples lack an explanation or discussion.
2. For the examples using real data, it is not clear where is the data located or how to obtain it.
3. In "FPCA of 1-dimensional sparse data": "Note that, here, we do not pass a dataset as argument of the transform method." -> Why?
4. Shouldn't the "Smoothing of dense two-dimensional functional data" and "Smoothing of dense one-dimensional functional data" examples be in the smoothing section?
5. In the "Representation of multivariate functional data" there is no mention of allowing functional data represented as a basis expansion.
6. The `PSplines` class seems not to have an associated documentation page.
