### 2024-12-24 StevenGolovkine/FDApy#26 @vnmabus
Issues with functionality/performance
A couple of issues with the functionality and performance, in relation with the open review in JOSS (openjournals/joss-reviews#7526).

1. Why is the FPCA functionality splitted in two classes, UFPCA and MFPCA? Wouldn't it be possible to have one FPCA and change the behavior depending on the type of the input?
2. It seems that irregular data with domain dimension >=2 still needs to be defined in a complete grid, just that the grid may be different for each function. That is, it is not possible to represent a function evaluated at discrete points in a surface. Is that interpretation right? Why was that chosen?
3. The implementation of irregular data using a separate array per observation seems to be difficult to scale performance-wise. Not only using separate arrays would be bad for the cache, but also it will be impossible to vectorize anything. Is that true in practice? Did you consider alternative implementations, and why did you choose this one?
4. It seems that the package right now is focused in representation of functional data and preprocessing, in particular dimensionality reduction. Are there plans to expand it to cover other areas, or what is planned as future work?
