### 2024-12-24 #7526 @vnmabus
I have submitted several issues in the target repository related with different aspects of the submission. I hope that the amount of observations do not discourage you, @StevenGolovkine. I tried to give all possible feedback so that you can improve your submission, but that does not mean that everything is of the same importance, or even that you need to do *all* the changes now. I also included some questions mostly out of personal curiosity.

I will update the issues when I receive your answers, and I could add a point of two if I notice that I missed something, but I do not expect to add many more substantial points.

I want to congratulate you again for the effort you put into this package, and wish you merry Christmas and a happy New Year.
