### 2024-12-27 StevenGolovkine/FDApy#27 @quantgirluk
Documentation 
- Documentation is for v1.0.0 but there are a couple of releases ahead on the repo. Is it possible to add details on any new functionality added, bugs fixed, backwards compatibility, etc. on each new version? This can also be added in the [main documentation](https://fdapy.readthedocs.io/en/stable/)
- In terms of Requirements, is there any particular version which is needed (e.g. package_name>=x.y.z or package_name<=x.y.z)?
