+++ joss/paper.md
@@ -25,24 +25,24 @@ Functional data analysis (FDA) is a statistical methodology for analyzing data t
 
 # Statement of need
 
-In order to apply FDA to real datasets, there is a need for appropriate softwares with up-to-date methodological implementation and easy addition of new theoretical developments. The seminal R package for FDA is \textsf{fda} [@ramsayFdaFunctionalData2023], based on work cited in @ramsayFunctionalDataAnalysis2005 and @ramsayFunctionalDataAnalysis2009. Most of the R packages that implements FDA methods are highly specialized and are built upon \textsf{fda}. For example, one may cite \textsf{FDboost} [@brockhausBoostingFunctionalRegression2020] and \textsf{refund} [@goldsmithRefundRegressionFunctional2023] for regression and classification, \textsf{funFEM} [@bouveyronFunFEMClusteringDiscriminative2021] and \textsf{funLBM} [@bouveyronFunLBMModelBasedCoClustering2022] for clustering or \textsf{fdasrvf} [@tuckerFdasrvfElasticFunctional2023] for functional data registration. For most packages, the functional data are however restricted to univariate functional data that are well described by their coefficients in a given basis of functions. The \textsf{funData} package [@happ-kurzObjectOrientedSoftwareFunctional2020] aims to provide a unified framework to handle univariate and multivariate functional data defined on different dimensional domains. Sparse functional data are also considered. The \textsf{MFPCA} [@happ-kurzMFPCAMultivariateFunctional2022] package, built on top of the \textsf{funData} package, implements multivariate functional principal components analysis (MFPCA) for data defined on different dimensional domains [@happMultivariateFunctionalPrincipal2018]. Consider looking at the CRAN webpage\footnote{\url{https://cran.r-project.org/web/views/FunctionalData.html}} on functional data to have a complete overview of the R packages.
+In order to apply FDA to real datasets, there is a need for appropriate softwares with up-to-date methodological implementation and easy addition of new theoretical developments. The seminal R package for FDA is `fda` [@ramsayFdaFunctionalData2023], based on work cited in @ramsayFunctionalDataAnalysis2005 and @ramsayFunctionalDataAnalysis2009. Most of the R packages that implement FDA methods are highly specialized and are built upon `fda`. For example, one may cite `FDboost` [@brockhausBoostingFunctionalRegression2020] and `refund` [@goldsmithRefundRegressionFunctional2023] for regression and classification, `funFEM` [@bouveyronFunFEMClusteringDiscriminative2021] and `funLBM` [@bouveyronFunLBMModelBasedCoClustering2022] for clustering or `fdasrvf` [@tuckerFdasrvfElasticFunctional2023] for functional data registration. For most packages, the functional data are however restricted to univariate functional data that are well described by their coefficients in a given basis of functions. The `funData` package [@happ-kurzObjectOrientedSoftwareFunctional2020] aims to provide a unified framework to handle univariate and multivariate functional data defined on different dimensional domains. Sparse functional data are also considered. The `MFPCA` [@happ-kurzMFPCAMultivariateFunctional2022] package, built on top of the `funData` package, implements multivariate functional principal components analysis (MFPCA) for data defined on different dimensional domains [@happMultivariateFunctionalPrincipal2018]. Consider looking at the CRAN webpage\footnote{\url{https://cran.r-project.org/web/views/FunctionalData.html}} on functional data to have a complete overview of the R packages.
 
[…]
