+++ joss/paper.bib
@@ -389,19 +389,20 @@
   doi = {10.1016/j.chemolab.2015.09.018}
 }
 
-@misc{ramos-carrenoScikitfdaPythonPackage2023,
+@article{ramos-carrenoScikitfdaPythonPackage2024,
   title = {Scikit-Fda: {{A Python Package}} for {{Functional Data Analysis}}},
   shorttitle = {Scikit-Fda},
-  author = {{Ramos-Carre{\~n}o}, Carlos and Torrecilla, Jos{\'e} Luis and {Carbajo-Berrocal}, Miguel and Marcos, Pablo and Su{\'a}rez, Alberto},
-  year = {2023},
-  month = jun,
-  number = {arXiv:2211.02566},
-  eprint = {2211.02566},
-  primaryclass = {cs, stat},
-  publisher = {arXiv},
-  doi = {10.48550/arXiv.2211.02566},
-  urldate = {2024-01-15},
+  author = {Ramos-Carreño, Carlos and Torrecilla, José Luis and Carbajo-Berrocal, Miguel and Marcos, Pablo and Suárez, Alberto},
+  date = {2024-05-08},
+  journaltitle = {Journal of Statistical Software},
+  volume = {109},
+  pages = {1--37},
+  issn = {1548-7660},
+  doi = {10.18637/jss.v109.i02},
+  url = {https://doi.org/10.18637/jss.v109.i02},
+  urldate = {2025-01-16},
 }
+
 
 @misc{loningSktimeSktimeV02022,
   title = {Sktime/Sktime: V0.13.4},
