+++ joss/paper.bib
@@ -460,3 +460,24 @@
   doi = {10.32614/cran.package.refund}
 }
 
+@article{yoshidaDetectingDifferencesGait2022,
+  title = {Detecting Differences in Gait Initiation between Older Adult Fallers and Non-Fallers through Multivariate Functional Principal Component Analysis},
+  author = {Yoshida, Kaya and Commandeur, Drew and Hundza, Sandra and Klimstra, Marc},
+  date = {2022-11-01},
+  journaltitle = {Journal of Biomechanics},
+  shortjournal = {Journal of Biomechanics},
+  volume = {144},
+  pages = {111342},
+  issn = {0021-9290},
+  doi = {10.1016/j.jbiomech.2022.111342},
+  url = {https://www.sciencedirect.com/science/article/pii/S0021929022003839},
+  urldate = {2025-01-17},
+}
+
+@thesis{nguyenLearningDomainspecificCameras2024,
+  title = {Learning Domain-Specific Cameras},
+  author = {Nguyen, Cindy My Anh},
+  date = {2024},
+  url = {https://purl.stanford.edu/zm136ny2176},
+  urldate = {2025-01-17},
+}
