### 2025-01-17 StevenGolovkine/FDApy#19 @StevenGolovkine
Hi @vnmabus,

I agree with your comments. I corrected the typos, added a general description of the package, changed the comment concerning `scikit-fda` and added references.

Let me know if it requires other changes.
