### 2025-01-19 StevenGolovkine/FDApy#28 @quantgirluk
Examples description
Hi @StevenGolovkine, thanks for developing this package and apologies for the late comments. I have been playing with the [examples](https://fdapy.readthedocs.io/en/latest/auto_examples/index.html) and have found them to work really well. My only comment is that from a user perspective it would be great to have more detailed explanations.

I would suggest to make the level of the details consistent across the examples. The examples in the [representation](https://fdapy.readthedocs.io/en/latest/auto_examples/representation/plot_dense_functional.html#sphx-glr-auto-examples-representation-plot-dense-functional-py) category are the ones which are easier to follow in my opinion. So, maybe you could make the rest of the examples consistent with these.
