### 2025-01-19 StevenGolovkine/FDApy#27 @quantgirluk
Re this sentence in the documentation: "This package provide modules for the analysis of such data. It includes methods for different dimensional data as well as irregularly sampled functional data. An implementation of (multivariate) functional principal component analysis is also given. Moreover, a simulation toolbox is provided. It might be used to simulate different clusters of functional data. "

- This package *provides
- Can you list the modules ("clustering",
        "misc",
        "preprocessing",
        "regression",
        "representation",
        "simulation",
        "visualization")
and provide a brief explanation of the functionality available on each one.
