### 2025-01-19 StevenGolovkine/FDApy#27 @quantgirluk
I would remove this sentence "Check out the documentation for more complete information on the available features within the package." from the documentation
