+++ docs/sg_execution_times.rst
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:27.102** total execution time for 28 files **from all galleries**:
+**00:23.004** total execution time for 28 files **from all galleries**:
 
 .. container::
 
@@ -33,86 +33,86 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_auto_examples_fpca_plot_mfpca_2d.py` (``../examples/fpca/plot_mfpca_2d.py``)
-     - 00:06.496
+     - 00:05.284
      - 0.0
    * - :ref:`sphx_glr_auto_examples_fpca_plot_mfpca_1d_2d.py` (``../examples/fpca/plot_mfpca_1d_2d.py``)
-     - 00:04.536
+     - 00:03.790
      - 0.0
    * - :ref:`sphx_glr_auto_examples_fpca_plot_mfpca_1d_sparse.py` (``../examples/fpca/plot_mfpca_1d_sparse.py``)
-     - 00:03.595
+     - 00:02.523
      - 0.0
    * - :ref:`sphx_glr_auto_examples_fpca_plot_mfpca_1d.py` (``../examples/fpca/plot_mfpca_1d.py``)
-     - 00:02.918
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_fpca_1d_sparse.py` (``../examples/fpca/plot_fpca_1d_sparse.py``)
-     - 00:02.398
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_fpca_2d.py` (``../examples/fpca/plot_fpca_2d.py``)
-     - 00:01.898
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_basis_plot_basis_2d.py` (``../examples/basis/plot_basis_2d.py``)
-     - 00:01.458
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_fpca_1d.py` (``../examples/fpca/plot_fpca_1d.py``)
-     - 00:00.958
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_simulation_plot_karhunen.py` (``../examples/simulation/plot_karhunen.py``)
-     - 00:00.618
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_simulation_plot_karhunen_multivariate.py` (``../examples/simulation/plot_karhunen_multivariate.py``)
-     - 00:00.594
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_basis_plot_basis_multivariate_1d.py` (``../examples/basis/plot_basis_multivariate_1d.py``)
-     - 00:00.499
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_simulation_plot_simulation.py` (``../examples/simulation/plot_simulation.py``)
-     - 00:00.357
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_simulation_plot_cluster_multivariate.py` (``../examples/simulation/plot_cluster_multivariate.py``)
-     - 00:00.170
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_basis_plot_basis_1d.py` (``../examples/basis/plot_basis_1d.py``)
-     - 00:00.168
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_representation_plot_smooth_data_2d.py` (``../examples/representation/plot_smooth_data_2d.py``)
-     - 00:00.140
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_basis_plot_basis_multivariate_2d.py` (``../examples/basis/plot_basis_multivariate_2d.py``)
-     - 00:00.116
-     - 0.0
[…]
