+++ pyproject.toml
@@ -46,7 +46,7 @@ test = [
   "pytest",
   "pytest-env",
   "pytest-subtests",
-  "scipy",
+  "scipy<1.11.0",
 ]
 
 [project.urls]
