### 2025-01-23 StevenGolovkine/FDApy#20 @StevenGolovkine
Hi @vnmabus,

I agree with your comments, I rewrote most of the main page and README.
Concerning Comment 4, at first, the software should have been compatible with scikit-learn (and that why it follows the fit/transform convention), but it is not the case anymore. From what I understood, scikit-learn requires tables (or similar) as entry and my representation of the data is not.

Let me know if it requires other changes.
