### 2025-01-23 StevenGolovkine/FDApy#20 @vnmabus
@StevenGolovkine I just want to note that, although most scikit-learn classes require NumPy arrays as input, others, like `GridSearch`, `Pipeline` and similar ones do not. That means that it is possible to use them with your own classes, and even have pipelines that combine functional and multivariate steps, as we do in scikit-fda. See https://fda.readthedocs.io/en/latest/auto_tutorial/plot_skfda_sklearn.html for an overview of what it works for us.
