### 2025-01-23 StevenGolovkine/FDApy#20 @StevenGolovkine
Thanks for the information, I was not aware of this possibility. For now, I'll stick with the non-compatibility and keep this in mind for eventual future work.
