### 2025-01-24 88f62c2
Create CODE_OF_CONDUCT.md
