+++ CONTRIBUTING.rst
@@ -66,7 +66,7 @@ Ready to contribute? Here's how to set up `FDApy` for local development.
 
     mkvirtualenv FDApy
     cd FDApy/
-    python setup.py develop
+    pip install -e .
 
 4. Create a branch for local development::
 
@@ -74,13 +74,12 @@ Ready to contribute? Here's how to set up `FDApy` for local development.
 
    Now you can make your changes locally.
 
-5. When you're done making changes, check that your changes pass flake8 and the tests, including testing other Python versions with tox::
+5. When you're done making changes, check that your changes pass flake8 and the tests::
 
     flake8 FDApy tests
-    python setup.py test or pytest
-    tox
+    pytest
 
-   To get flake8 and tox, just pip install them into your virtualenv.
+   To get flake8, just pip install it into your virtualenv.
 
 6. Commit your changes and push your branch to GitHub::
 
