+++ CONTRIBUTING.rst
@@ -98,7 +98,7 @@ Before you submit a pull request, check that it meets these guidelines:
 2. If the pull request adds functionality, the docs should be updated. Put
    your new functionality into a function with a docstring, and add the
    feature to the list in docs/index.rst.
-3. The pull request should work for Python 3.9 and for PyPI.
+3. The pull request should work for Python 3.10 and above.
 
 Tips
 ----
