### 2025-01-25 StevenGolovkine/FDApy#19 @vnmabus
Some additional typos on a second read:

"FDA has been successfully applied in different context" -> "FDA has been successfully applied in different contexts"
"a library developted" -> " a library developed"

Apart from that, it looks good to me.
