### 2025-01-25 StevenGolovkine/FDApy#20 @vnmabus
Thanks for taking my feedback into consideration. I do not think I have additional feedback for this part of the docs. so feel free to close this issue.
