### 2025-01-25 StevenGolovkine/FDApy#19 @StevenGolovkine
Typo fixed. Thanks for your careful reading.
I'm closing the issue now.
