### 2025-01-26 StevenGolovkine/FDApy#21 @StevenGolovkine
Hi @vnmabus,

Thanks for your comments. I changed the outdated explanations and included files to make things clearer.
For comment 5, actually I'm not sure of what that means, so I changed it by: "The pull request should work for Python 3.10 and above."
