+++ docs/modules/simulation.rst
@@ -1,20 +1,20 @@
 Simulation
 ==========
 
-.. toctree::
-	:includehidden:
-	:maxdepth: 2
-	:titlesonly:
-
-This module is used to define an abstract Simulation class. We may simulate
-different data from a linear combination of basis functions or multiple
-realizations of diverse Brownian motion.
-
-.. automodule:: FDApy.simulation.simulation
-    :members:
-
-.. automodule:: FDApy.simulation.brownian
-    :members:
-
-.. automodule:: FDApy.simulation.karhunen
-    :members:
+.. .. toctree::
+.. 	:includehidden:
+.. 	:maxdepth: 2
+.. 	:titlesonly:
+.. 
+.. This module is used to define an abstract Simulation class. We may simulate
+.. different data from a linear combination of basis functions or multiple
+.. realizations of diverse Brownian motion.
+.. 
+.. .. automodule:: FDApy.simulation.simulation
+..     :members:
+.. 
+.. .. automodule:: FDApy.simulation.brownian
+..     :members:
+.. 
+.. .. automodule:: FDApy.simulation.karhunen
+..     :members:
