+++ docs/modules/misc.rst
@@ -1,13 +1,13 @@
 Miscellaneous
 =============
 
-.. toctree::
-	:includehidden:
-	:maxdepth: 2
-	:titlesonly:
-
-.. automodule:: FDApy.misc.loader
-    :members:
-
-.. automodule:: FDApy.misc.utils
-    :members:
+.. .. toctree::
+.. 	:includehidden:
+.. 	:maxdepth: 2
+.. 	:titlesonly:
+.. 
+.. .. automodule:: FDApy.misc.loader
+..     :members:
+.. 
+.. .. automodule:: FDApy.misc.utils
+..     :members:
