+++ docs/modules/autosummary/FDApy.representation.MultivariateFunctionalData.rst
@@ -1,0 +1,171 @@
+﻿MultivariateFunctionalData
+==========================
+
+.. currentmodule:: FDApy.representation
+
+.. autoclass:: MultivariateFunctionalData
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+      
+      ~MultivariateFunctionalData.append
+      
+      
+      ~MultivariateFunctionalData.center
+      
+      
+      ~MultivariateFunctionalData.clear
+      
+      
+      ~MultivariateFunctionalData.concatenate
+      
+      
+      ~MultivariateFunctionalData.copy
+      
+      
+      ~MultivariateFunctionalData.count
+      
+      
+      ~MultivariateFunctionalData.covariance
+      
+      
+      ~MultivariateFunctionalData.extend
+      
+      
+      ~MultivariateFunctionalData.index
+      
+      
+      ~MultivariateFunctionalData.inner_product
+      
+      
+      ~MultivariateFunctionalData.insert
+      
+      
+      ~MultivariateFunctionalData.mean
+      
+      
+      ~MultivariateFunctionalData.noise_variance
+      
+      
+      ~MultivariateFunctionalData.norm
+      
+      
+      ~MultivariateFunctionalData.normalize
+      
+      
+      ~MultivariateFunctionalData.pop
+      
+      
+      ~MultivariateFunctionalData.remove
+      
+      
+      ~MultivariateFunctionalData.rescale
+      
+      
+      ~MultivariateFunctionalData.reverse
+      
+      
+      ~MultivariateFunctionalData.smooth
+      
+      
+      ~MultivariateFunctionalData.sort
+      
+      
+      ~MultivariateFunctionalData.standardize
+      
+      
+      ~MultivariateFunctionalData.to_basis
+      
+      
+      ~MultivariateFunctionalData.to_grid
+      
+      
+      ~MultivariateFunctionalData.to_long
+      
+   
+   
+   
+      
+      
+   .. automethod:: append
+      
+      
+   .. automethod:: center
+      
+      
+   .. automethod:: clear
+      
+      
+   .. automethod:: concatenate
+      
+      
+   .. automethod:: copy
+      
+      
+   .. automethod:: count
+      
+      
+   .. automethod:: covariance
+      
+      
+   .. automethod:: extend
+      
+      
+   .. automethod:: index
+      
+      
+   .. automethod:: inner_product
+      
+      
+   .. automethod:: insert
+      
+      
+   .. automethod:: mean
+      
+      
+   .. automethod:: noise_variance
+      
+      
+   .. automethod:: norm
+      
+      
+   .. automethod:: normalize
+      
+      
+   .. automethod:: pop
+      
+      
+   .. automethod:: remove
+      
+      
+   .. automethod:: rescale
+      
+      
+   .. automethod:: reverse
+      
+      
+   .. automethod:: smooth
+      
+      
+   .. automethod:: sort
+      
+      
+   .. automethod:: standardize
+      
+      
+   .. automethod:: to_basis
+      
+      
+   .. automethod:: to_grid
+      
+      
+   .. automethod:: to_long
+      
+   
+   
+.. include:: /backreferences/FDApy.representation.MultivariateFunctionalData.examples
