+++ docs/modules/autosummary/FDApy.representation.IrregularFunctionalData.rst
@@ -1,0 +1,99 @@
+﻿IrregularFunctionalData
+=======================
+
+.. currentmodule:: FDApy.representation
+
+.. autoclass:: IrregularFunctionalData
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+      
+      ~IrregularFunctionalData.center
+      
+      
+      ~IrregularFunctionalData.concatenate
+      
+      
+      ~IrregularFunctionalData.covariance
+      
+      
+      ~IrregularFunctionalData.inner_product
+      
+      
+      ~IrregularFunctionalData.mean
+      
+      
+      ~IrregularFunctionalData.noise_variance
+      
+      
+      ~IrregularFunctionalData.norm
+      
+      
+      ~IrregularFunctionalData.normalize
+      
+      
+      ~IrregularFunctionalData.rescale
+      
+      
+      ~IrregularFunctionalData.smooth
+      
+      
+      ~IrregularFunctionalData.standardize
+      
+      
+      ~IrregularFunctionalData.to_basis
+      
+      
+      ~IrregularFunctionalData.to_long
+      
+   
+   
+   
+      
+      
+   .. automethod:: center
+      
+      
+   .. automethod:: concatenate
+      
+      
+   .. automethod:: covariance
+      
+      
+   .. automethod:: inner_product
+      
+      
+   .. automethod:: mean
+      
+      
+   .. automethod:: noise_variance
+      
+      
+   .. automethod:: norm
+      
+      
+   .. automethod:: normalize
+      
+      
+   .. automethod:: rescale
+      
+      
+   .. automethod:: smooth
+      
+      
+   .. automethod:: standardize
+      
+      
+   .. automethod:: to_basis
+      
+      
+   .. automethod:: to_long
+      
+   
+   
+.. include:: /backreferences/FDApy.representation.IrregularFunctionalData.examples
