### 2025-01-27 91ec913
References functional data
