+++ FDApy/representation/functional_data.py
@@ -204,7 +204,7 @@ def _estimate_noise_variance_with_covariance(
 
     References
     ----------
-    .. [2] Yao, F., Müller, H.-G., Wang, J.-L. (2005). Functional Data
+    .. [3] Yao, F., Müller, H.-G., Wang, J.-L. (2005). Functional Data
         Analysis for Sparse Longitudinal Data. Journal of the American
         Statistical Association 100, pp. 577--590.
 
@@ -835,6 +835,31 @@ class DenseFunctionalData(GridFunctionalData):
     ... ]))
     >>> DenseFunctionalData(argvals, values)
 
+    References
+    ----------
+    .. [1] Benko, M., Härdle, W. and Kneip, A. (2009). Common functional
+        principal components. The Annals of Statistics 37, 1--34.
+    .. [2] Cai, T.T., Yuan, M., (2011), Optimal estimation of the mean
+        function based on discretely sampled functional data: Phase
+        transition. The Annals of Statistics 39, 2330-2355.
+    .. [3] Chiou, J.-M., Chen, Y.-T., Yang, Y.-F. (2014). Multivariate Functional
+        Principal Component Analysis: A Normalization Approach. Statistica Sinica
+        24, 1571--1596.
+    .. [4] Eilers, P. H. C., Marx, B. D. (2021). Practical Smoothing: The Joys of
+        P-splines. Cambridge University Press, Cambridge.
+    .. [5] Hall, P., Kay, J.W. and Titterington, D.M. (1990).
+        Asymptotically Optimal Difference-Based Estimation of Variance in
+        Nonparametric Regression. Biometrika 77, 521--528.
+    .. [6] Happ, C., Greven, S. (2018). Multivariate Functional Principal Component
+        Analysis for Data Observed on Different (Dimensional) Domains. Journal of
+        the American Statistical Association 113, 649--659.
+    .. [7] Ramsey, J. O. and Silverman, B. W. (2005), Functional Data
+        Analysis, Springer Science, Chapter 8.
+    .. [8] Tsybakov, A.B. (2008), Introduction to Nonparametric Estimation.
+        Springer Series in Statistics.
+    .. [9] Zhang, J.-T. and Chen J. (2007), Statistical Inferences for
+        Functional Data, The Annals of Statistics, Vol. 35, No. 3.
+
     """
 
     ###########################################################################
@@ -1089,18 +1114,12 @@ class DenseFunctionalData(GridFunctionalData):
         ----------
         order: int, default=2
             Order of the difference sequence. The order has to be between
-            1 and 10. See [1]_ for more information.
+            1 and 10. See [4]_ for more information.
 
         Returns
         -------
         float
             The estimation of the variance of the noise.
-
-        References
-        ----------
-        .. [1] Hall, P., Kay, J.W. and Titterington, D.M. (1990).
-            Asymptotically Optimal Difference-Based Estimation of Variance in
-            Nonparametric Regression. Biometrika 77, 521--528.
 
         Examples
         --------
@@ -1148,13 +1167,13 @@ class DenseFunctionalData(GridFunctionalData):
             Points at which the curves are estimated. The default is None,
             meaning we use the argvals as estimation points.
         method: str, default='PS'
-            The method to used for the smoothing. If 'PS', the method is P-splines [1]_.
-            If 'LP', the method is local polynomials [3]_. Otherwise, it raises an
+            The method to used for the smoothing. If 'PS', the method is P-splines [2]_.
+            If 'LP', the method is local polynomials [4]_. Otherwise, it raises an
             error.
         bandwidth: Optional[float], default=None
[…]
