+++ docs/modules/autosummary/FDApy.representation.functional_data.DenseFunctionalData.rst
@@ -1,99 +1,0 @@
-﻿DenseFunctionalData
-===================
-
-.. currentmodule:: FDApy.representation.functional_data
-
-.. autoclass:: DenseFunctionalData
-
-   
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      
-      
-      ~DenseFunctionalData.center
-      
-      
-      ~DenseFunctionalData.concatenate
-      
-      
-      ~DenseFunctionalData.covariance
-      
-      
-      ~DenseFunctionalData.inner_product
-      
-      
-      ~DenseFunctionalData.mean
-      
-      
-      ~DenseFunctionalData.noise_variance
-      
-      
-      ~DenseFunctionalData.norm
-      
-      
-      ~DenseFunctionalData.normalize
-      
-      
-      ~DenseFunctionalData.rescale
-      
-      
-      ~DenseFunctionalData.smooth
-      
-      
-      ~DenseFunctionalData.standardize
-      
-      
-      ~DenseFunctionalData.to_basis
-      
-      
-      ~DenseFunctionalData.to_long
-      
-   
-   
-   
-      
-      
-   .. automethod:: center
-      
-      
-   .. automethod:: concatenate
-      
-      
-   .. automethod:: covariance
-      
-      
-   .. automethod:: inner_product
-      
-      
-   .. automethod:: mean
-      
-      
-   .. automethod:: noise_variance
-      
-      
-   .. automethod:: norm
-      
-      
-   .. automethod:: normalize
-      
-      
-   .. automethod:: rescale
-      
-      
-   .. automethod:: smooth
-      
-      
-   .. automethod:: standardize
-      
-      
-   .. automethod:: to_basis
-      
-      
-   .. automethod:: to_long
-      
-   
-   
-.. include:: /backreferences/FDApy.representation.functional_data.DenseFunctionalData.examples
