+++ docs/modules/autosummary/FDApy.representation.DenseFunctionalDataIterator.rst
@@ -1,0 +1,21 @@
+﻿DenseFunctionalDataIterator
+===========================
+
+.. currentmodule:: FDApy.representation
+
+.. autoclass:: DenseFunctionalDataIterator
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+   
+   
+   
+      
+   
+   
+.. include:: /backreferences/FDApy.representation.DenseFunctionalDataIterator.examples
