+++ docs/modules/autosummary/FDApy.representation.BasisFunctionalDataIterator.rst
@@ -1,0 +1,21 @@
+﻿BasisFunctionalDataIterator
+===========================
+
+.. currentmodule:: FDApy.representation
+
+.. autoclass:: BasisFunctionalDataIterator
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+   
+   
+   
+      
+   
+   
+.. include:: /backreferences/FDApy.representation.BasisFunctionalDataIterator.examples
