### 2025-01-28 d2ab4c8
References functional_data
