+++ FDApy/representation/functional_data.py
@@ -118,8 +118,8 @@ def _smooth_covariance(
 
     References
     ----------
-    .. [1] Eilers, P. H. C., Marx, B. D. (2021). Practical Smoothing: The Joys of
-        P-splines. Cambridge University Press, Cambridge.
+    .. [1] Eilers, P. H. C., Marx, B. D. (2021). Practical Smoothing: The Joys
+        of P-splines. Cambridge University Press, Cambridge.
     .. [2] Zhang, J.-T. and Chen J. (2007), Statistical Inferences for
         Functional Data, The Annals of Statistics, Vol. 35, No. 3.
 
@@ -204,7 +204,7 @@ def _estimate_noise_variance_with_covariance(
 
     References
     ----------
-    .. [3] Yao, F., Müller, H.-G., Wang, J.-L. (2005). Functional Data
+    .. [1] Yao, F., Müller, H.-G., Wang, J.-L. (2005). Functional Data
         Analysis for Sparse Longitudinal Data. Journal of the American
         Statistical Association 100, pp. 577--590.
 
@@ -781,14 +781,15 @@ class DenseFunctionalData(GridFunctionalData):
     A class used to define dense functional data. We denote by :math:`n`, the
     number of observations and by :math:`p`, the number of input dimensions.
     Here, we are in the case of univariate functional data, and so the output
-    dimension will be :math:`\mathbb{R}`. We note by :math:`X` an observation, while
-    we use :math:`X_1, \dots, X_n` if we refer to a particular set of observations. The
-    observations are defined as:
+    dimension will be :math:`\mathbb{R}`. We note by :math:`X` an observation,
+    while we use :math:`X_1, \dots, X_n` if we refer to a particular set of
+    observations. The observations are defined as:
 
     .. math::
         X(t): \mathcal{T} \longrightarrow \mathbb{R},
 
-    where :math:`\mathcal{T} \subset \mathbb{R}^p`. We denote the mean function by
+    where :math:`\mathcal{T} \subset \mathbb{R}^p`. We denote the mean function
+    by
 
     .. math::
         \mu(t): \mathcal{T} \longrightarrow \mathbb{R},
@@ -842,23 +843,21 @@ class DenseFunctionalData(GridFunctionalData):
     .. [2] Cai, T.T., Yuan, M., (2011), Optimal estimation of the mean
         function based on discretely sampled functional data: Phase
         transition. The Annals of Statistics 39, 2330-2355.
-    .. [3] Chiou, J.-M., Chen, Y.-T., Yang, Y.-F. (2014). Multivariate Functional
-        Principal Component Analysis: A Normalization Approach. Statistica Sinica
-        24, 1571--1596.
-    .. [4] Eilers, P. H. C., Marx, B. D. (2021). Practical Smoothing: The Joys of
-        P-splines. Cambridge University Press, Cambridge.
+    .. [3] Chiou, J.-M., Chen, Y.-T., Yang, Y.-F. (2014). Multivariate
+        Functional Principal Component Analysis: A Normalization Approach.
+        Statistica Sinica 24, 1571--1596.
+    .. [4] Eilers, P. H. C., Marx, B. D. (2021). Practical Smoothing: The Joys
+        of P-splines. Cambridge University Press, Cambridge.
     .. [5] Hall, P., Kay, J.W. and Titterington, D.M. (1990).
         Asymptotically Optimal Difference-Based Estimation of Variance in
         Nonparametric Regression. Biometrika 77, 521--528.
-    .. [6] Happ, C., Greven, S. (2018). Multivariate Functional Principal Component
-        Analysis for Data Observed on Different (Dimensional) Domains. Journal of
-        the American Statistical Association 113, 649--659.
[…]
