+++ setup.cfg
@@ -18,7 +18,7 @@ ignore_missing_imports = True
 ignore_missing_imports = True
 
 [flake8]
-max-line-length = 88
+max-line-length = 80
 extend-ignore = E203
 
 [coverage:run]
