+++ FDApy/preprocessing/dim_reduction/__init__.py
@@ -1,4 +1,5 @@
 """Dimension reduction module."""
+
 import lazy_loader as lazy
 
 __getattr__, __dir__, __all__ = lazy.attach(
