+++ FDApy/representation/functional_data.py
@@ -1347,16 +1347,11 @@ class DenseFunctionalData(GridFunctionalData):
         if mean is None:
             data_mean = self.mean(method_smoothing=method_smoothing, **kwargs)
         elif (mean is not None) and (method_smoothing is not None):
-            data_mean = mean.smooth(
-                self.argvals,
-                method=method_smoothing,
-                **kwargs
-            )
+            data_mean = mean.smooth(self.argvals, method=method_smoothing, **kwargs)
         else:
             data_mean = mean
         return DenseFunctionalData(
-            DenseArgvals(self.argvals),
-            DenseValues(self.values - data_mean.values)
+            DenseArgvals(self.argvals), DenseValues(self.values - data_mean.values)
         )
 
     def norm(
@@ -1414,11 +1409,7 @@ class DenseFunctionalData(GridFunctionalData):
 
         norm_fd = np.zeros(n_obs)
         for idx in np.arange(n_obs):
-            norm_fd[idx] = _integrate(
-                sq_values[idx],
-                *axis,
-                method=method_integration
-            )
+            norm_fd[idx] = _integrate(sq_values[idx], *axis, method=method_integration)
 
         if squared:
             return np.array(norm_fd)
@@ -1649,10 +1640,7 @@ class DenseFunctionalData(GridFunctionalData):
         for i, j in itertools.product(np.arange(n_obs), repeat=2):
             if i <= j:
                 inner_mat[i, j] = _inner_product(
-                    data.values[i],
-                    data.values[j],
-                    *axis,
-                    method=method_integration
+                    data.values[i], data.values[j], *axis, method=method_integration
                 )
         inner_mat = inner_mat - np.diag(np.repeat(self._noise_variance, n_obs))
 
@@ -1719,9 +1707,7 @@ class DenseFunctionalData(GridFunctionalData):
 
         """
         if self.n_dimension > 1:
-            raise ValueError(
-                "Only one dimensional functional data are supported."
-            )
+            raise ValueError("Only one dimensional functional data are supported.")
 
         if points is None:
             points = self.argvals
@@ -1741,9 +1727,7 @@ class DenseFunctionalData(GridFunctionalData):
         # Center the data
         data = self
         if center:
-            data = data.center(
-                method_smoothing=method_smoothing, **kwargs_center
-            )
+            data = data.center(method_smoothing=method_smoothing, **kwargs_center)
 
         # Estimate the covariance
         cov = np.dot(data.values.T, data.values) / (self.n_obs - 1)
@@ -1769,10 +1753,7 @@ class DenseFunctionalData(GridFunctionalData):
             raw_diag_cov, np.diag(cov), self.argvals, points
         )
 
-        self._covariance = DenseFunctionalData(
-            points_cov,
-            DenseValues(cov[np.newaxis])
-        )
+        self._covariance = DenseFunctionalData(points_cov, DenseValues(cov[np.newaxis]))
         return self._covariance
 
     ###########################################################################
@@ -1893,9 +1874,7 @@ class IrregularFunctionalData(GridFunctionalData):
     # Static methods
     @staticmethod
     def _perform_computation(
-        fdata1: IrregularFunctionalData,
-        fdata2: IrregularFunctionalData,
-        func: Callable
+        fdata1: IrregularFunctionalData, fdata2: IrregularFunctionalData, func: Callable
     ) -> IrregularFunctionalData:
         """Perform computation defined by `func` if they are compatible.
 
@@ -1922,10 +1901,7 @@ class IrregularFunctionalData(GridFunctionalData):
                 fdata1.values.items(), fdata2.values.items()
             )
         }
-        return IrregularFunctionalData(
-            fdata1.argvals,
[…]
