+++ docs/modules/autosummary/FDApy.preprocessing.LocalPolynomial.rst
@@ -1,0 +1,27 @@
+﻿LocalPolynomial
+===============
+
+.. currentmodule:: FDApy.preprocessing
+
+.. autoclass:: LocalPolynomial
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+      
+      ~LocalPolynomial.predict
+      
+   
+   
+   
+      
+      
+   .. automethod:: predict
+      
+   
+   
+.. include:: /backreferences/FDApy.preprocessing.LocalPolynomial.examples
