+++ docs/modules/autosummary/FDApy.preprocessing.MFPCA.rst
@@ -1,0 +1,39 @@
+﻿MFPCA
+=====
+
+.. currentmodule:: FDApy.preprocessing
+
+.. autoclass:: MFPCA
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+      
+      ~MFPCA.fit
+      
+      
+      ~MFPCA.inverse_transform
+      
+      
+      ~MFPCA.transform
+      
+   
+   
+   
+      
+      
+   .. automethod:: fit
+      
+      
+   .. automethod:: inverse_transform
+      
+      
+   .. automethod:: transform
+      
+   
+   
+.. include:: /backreferences/FDApy.preprocessing.MFPCA.examples
