+++ docs/modules/autosummary/FDApy.preprocessing.FCPTPA.rst
@@ -1,0 +1,39 @@
+﻿FCPTPA
+======
+
+.. currentmodule:: FDApy.preprocessing
+
+.. autoclass:: FCPTPA
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+      
+      ~FCPTPA.fit
+      
+      
+      ~FCPTPA.inverse_transform
+      
+      
+      ~FCPTPA.transform
+      
+   
+   
+   
+      
+      
+   .. automethod:: fit
+      
+      
+   .. automethod:: inverse_transform
+      
+      
+   .. automethod:: transform
+      
+   
+   
+.. include:: /backreferences/FDApy.preprocessing.FCPTPA.examples
