+++ FDApy/preprocessing/dim_reduction/mfpca.py
@@ -359,7 +359,7 @@ class MFPCA:
         eigendecomposition of the covariance operator of each univariate
         components. If `method == 'inner-product'`, the estimation is
         based on an eigendecomposition of the inner-product matrix.
-    weights: npt.NDArray[np.float_], default=None
+    weights: npt.NDArray[np.float64], default=None
         A vector of weights of length :math:`P`. If `None`, we set the weights to be
         equal to 1 for each component.
     normalize: bool, default=False
@@ -372,7 +372,7 @@ class MFPCA:
     covariance: MultivariateFunctionalData
         An estimation of the covariance of the training data based on their
         eigendecomposition using the Mercer's theorem.
-    eigenvalues: npt.NDArray[np.float_], shape=(n_components,)
+    eigenvalues: npt.NDArray[np.float64], shape=(n_components,)
         The singular values corresponding to each of selected components.
     eigenfunctions: MultivariateFunctionalData
         Principal axes in feature space, representing the directions of
@@ -380,7 +380,10 @@ class MFPCA:
 
     References
     ----------
-    .. [1] Happ and Greven (2018), Multivariate Functional Principal
+    .. [1] Golovkine, S., Gunning, E., Simpkin, A.J., Bargary, N. (2023). On the use
+        of the Gram matrix for multivariate functional principal components
+        analysis.
+    .. [2] Happ and Greven (2018), Multivariate Functional Principal
         Component Analysis for Data Observed on Different (Dimensional)
         Domains. Journal of the American Statistical Association, 113,
         pp. 649--659.
@@ -392,7 +395,7 @@ class MFPCA:
         n_components: Union[int, float] = 2,
         univariate_expansions: Optional[List[Dict[str, object]]] = None,
         method: str = "covariance",
-        weights: Optional[npt.NDArray[np.float_]] = None,
+        weights: Optional[npt.NDArray[np.float64]] = None,
         normalize: bool = False,
     ) -> None:
         """Initialize MFPCA object."""
@@ -432,12 +435,12 @@ class MFPCA:
         self._method = new_method
 
     @property
-    def weights(self) -> Optional[npt.NDArray[np.float_]]:
+    def weights(self) -> Optional[npt.NDArray[np.float64]]:
         """Getter for `weights`."""
         return self._weights
 
     @weights.setter
-    def weights(self, new_weights: Optional[npt.NDArray[np.float_]]) -> None:
+    def weights(self, new_weights: Optional[npt.NDArray[np.float64]]) -> None:
         self._weights = new_weights
 
     @property
@@ -460,7 +463,7 @@ class MFPCA:
         return self._covariance
 
     @property
-    def eigenvalues(self) -> npt.NDArray[np.float_]:
+    def eigenvalues(self) -> npt.NDArray[np.float64]:
         """Getter for `eigenvalues`."""
         return self._eigenvalues
 
@@ -496,16 +499,6 @@ class MFPCA:
 
             - :meth:`FunctionalData.mean` and :meth:`FunctionalData.center`;
             - :meth:`preprocessing.dim_reduction.mfpca._fit_inner_product_multivariate`.
-
-        References
-        ----------
-        .. [1] Golovkine, S., Gunning, E., Simpkin, A.J., Bargary, N. (2023). On the use
-            of the Gram matrix for multivariate functional principal components
-            analysis.
-        .. [2] Happ C. & Greven S. (2018), Multivariate Functional Principal
-            Component Analysis for Data Observed on Different (Dimensional)
-            Domains. Journal of the American Statistical Association, 113,
-            pp. 649--659.
 
         """
         if points is None:
@@ -563,7 +556,6 @@ class MFPCA:
         self._basis_univariate = results.get("_basis_univariate", None)
         self._training_data = data
 
[…]
