+++ FDApy/preprocessing/dim_reduction/ufpca.py
@@ -68,12 +68,6 @@ def _fit_covariance(
         - `'eigenfunctions'`: the estimated eigenfunctions;
         - `'noise_variance_cov'`: the estimation of the noise.
 
-
-    References
-    ----------
-    .. [1] Ramsey, J. O. and Silverman, B. W. (2005), Functional Data
-        Analysis, Springer Science, Chapter 8.
-
     """
     # Compute the covariance
     covariance = data.covariance(
@@ -207,8 +201,6 @@ def _transform_numerical_integration_irregular(
     method: str = "trapz",
 ) -> npt.NDArray[np.float64]:
     """Estimate scores using numerical integration.
-
-    TODO: Consider another way to choose the bandwidth.
 
     Parameters
     ----------
@@ -416,6 +408,10 @@ class UFPCA:
     ----------
     .. [1] Ramsey, J. O. and Silverman, B. W. (2005), Functional Data
         Analysis, Springer Science, Chapter 8.
+    .. [2] Yao, Müller and Wang (2005), Functional Data Analysis for Sparse
+        Longitudinal Data. Journal of the American Statistical Association,
+        100, pp. 577--590.
+
 
     """
 
@@ -509,11 +505,6 @@ class UFPCA:
         kwargs_innpro: Dict[str, object], default={}
             Keywords arguments to be passed to the function
             :meth:`preprocessing.fpca._fit_inner_product`.
-
-        References
-        ----------
-        .. [1] Ramsey, J. O. and Silverman, B. W. (2005), Functional Data
-            Analysis, Springer Science, Chapter 8.
 
         """
         if self.method == "covariance" and data.n_dimension > 1:
@@ -621,7 +612,7 @@ class UFPCA:
         This integral can be estimated using two ways. First, if data are
         sampled on a common fine grid, the estimation is done using
         numerical integration. Second, the PACE (Principal Components through
-        Conditional Expectation) algorithm [1]_ is used for sparse functional
+        Conditional Expectation) algorithm [2]_ is used for sparse functional
         data. If the eigenfunctions have been estimated using the inner-product
         matrix, the scores can also be estimated using the formula
 
@@ -659,12 +650,6 @@ class UFPCA:
         npt.NDArray[np.float64], shape=(n_obs, n_components)
             An array representing the projection of the data onto the basis of
             functions defined by the eigenfunctions.
-
-        References
-        ----------
-        .. [1] Yao, Müller and Wang (2005), Functional Data Analysis for Sparse
-            Longitudinal Data. Journal of the American Statistical Association,
-            100, pp. 577--590.
 
         """
         # Checkers
