+++ docs/modules/autosummary/FDApy.simulation.Simulation.rst
@@ -1,0 +1,45 @@
+﻿Simulation
+==========
+
+.. currentmodule:: FDApy.simulation
+
+.. autoclass:: Simulation
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+      
+      ~Simulation.add_noise
+      
+      
+      ~Simulation.add_noise_and_sparsify
+      
+      
+      ~Simulation.new
+      
+      
+      ~Simulation.sparsify
+      
+   
+   
+   
+      
+      
+   .. automethod:: add_noise
+      
+      
+   .. automethod:: add_noise_and_sparsify
+      
+      
+   .. automethod:: new
+      
+      
+   .. automethod:: sparsify
+      
+   
+   
+.. include:: /backreferences/FDApy.simulation.Simulation.examples
