+++ docs/modules/visualization.rst
@@ -1,10 +1,0 @@
-Visualization
-=============
-
-.. .. toctree::
-.. 	:includehidden:
-.. 	:maxdepth: 2
-.. 	:titlesonly:
-.. 
-.. .. automodule:: FDApy.visualization.plot
-..     :members:
