+++ docs/modules/autosummary/FDApy.visualization.plot.rst
@@ -1,0 +1,8 @@
+﻿plot
+====
+
+.. currentmodule:: FDApy.visualization
+
+.. autofunction:: plot
+
+.. include:: /backreferences/FDApy.visualization.plot.examples
