+++ docs/modules/autosummary/FDApy.representation.IrregularValues.rst
@@ -1,0 +1,99 @@
+﻿IrregularValues
+===============
+
+.. currentmodule:: FDApy.representation
+
+.. autoclass:: IrregularValues
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+      
+      ~IrregularValues.clear
+      
+      
+      ~IrregularValues.compatible_with
+      
+      
+      ~IrregularValues.concatenate
+      
+      
+      ~IrregularValues.copy
+      
+      
+      ~IrregularValues.fromkeys
+      
+      
+      ~IrregularValues.get
+      
+      
+      ~IrregularValues.items
+      
+      
+      ~IrregularValues.keys
+      
+      
+      ~IrregularValues.pop
+      
+      
+      ~IrregularValues.popitem
+      
+      
+      ~IrregularValues.setdefault
+      
+      
+      ~IrregularValues.update
+      
+      
+      ~IrregularValues.values
+      
+   
+   
+   
+      
+      
+   .. automethod:: clear
+      
+      
+   .. automethod:: compatible_with
+      
+      
+   .. automethod:: concatenate
+      
+      
+   .. automethod:: copy
+      
+      
+   .. automethod:: fromkeys
+      
+      
+   .. automethod:: get
+      
+      
+   .. automethod:: items
+      
+      
+   .. automethod:: keys
+      
+      
+   .. automethod:: pop
+      
+      
+   .. automethod:: popitem
+      
+      
+   .. automethod:: setdefault
+      
+      
+   .. automethod:: update
+      
+      
+   .. automethod:: values
+      
+   
+   
+.. include:: /backreferences/FDApy.representation.IrregularValues.examples
