+++ docs/modules/autosummary/FDApy.representation.IrregularArgvals.rst
@@ -1,0 +1,117 @@
+﻿IrregularArgvals
+================
+
+.. currentmodule:: FDApy.representation
+
+.. autoclass:: IrregularArgvals
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      
+      
+      ~IrregularArgvals.clear
+      
+      
+      ~IrregularArgvals.compatible_with
+      
+      
+      ~IrregularArgvals.concatenate
+      
+      
+      ~IrregularArgvals.copy
+      
+      
+      ~IrregularArgvals.fromkeys
+      
+      
+      ~IrregularArgvals.get
+      
+      
+      ~IrregularArgvals.items
+      
+      
+      ~IrregularArgvals.keys
+      
+      
+      ~IrregularArgvals.normalization
+      
+      
+      ~IrregularArgvals.pop
+      
+      
+      ~IrregularArgvals.popitem
+      
+      
+      ~IrregularArgvals.setdefault
+      
+      
+      ~IrregularArgvals.switch
+      
+      
+      ~IrregularArgvals.to_dense
+      
+      
+      ~IrregularArgvals.update
+      
+      
+      ~IrregularArgvals.values
+      
+   
+   
+   
+      
+      
+   .. automethod:: clear
+      
+      
+   .. automethod:: compatible_with
+      
+      
+   .. automethod:: concatenate
+      
+      
+   .. automethod:: copy
+      
+      
+   .. automethod:: fromkeys
+      
+      
+   .. automethod:: get
+      
+      
+   .. automethod:: items
+      
+      
+   .. automethod:: keys
+      
+      
+   .. automethod:: normalization
+      
+      
+   .. automethod:: pop
+      
+      
+   .. automethod:: popitem
+      
+      
+   .. automethod:: setdefault
+      
+      
+   .. automethod:: switch
+      
+      
+   .. automethod:: to_dense
+      
+      
+   .. automethod:: update
+      
+      
+   .. automethod:: values
+      
+   
+   
+.. include:: /backreferences/FDApy.representation.IrregularArgvals.examples
