+++ docs/modules/representation.rst
@@ -8,7 +8,7 @@ The first step is the analysis of functional data is to represent them in a suit
 Generic representation
 ======================
 
-Representations of functional data are based on three abstract classes. The first one, :class:`Argvals`, represents the arguments of the functions. The second one, :class:`Values`, represents the values of the functions. The third one, :class:`FunctionalData`, represents the functional data object, which is a pair of :class:`Argvals` and :class:`Values`. The package provides several implementations of these classes, depending on the type of functional data to represent.
+Representations of functional data are based on three abstract classes. The first one, :class:`Argvals`, represents the arguments of the functions. The second one, :class:`Values`, represents the values of the functions. The third one, :class:`FunctionalData`, represents the functional data object, which is a pair of :class:`Argvals` and :class:`Values`. The package provides several implementations of these classes, depending on the type of functional data to represent. These classes cannot be instantiated directly, but they are used as base classes for the specific implementations.
 
 .. autosummary::
 	:toctree: autosummary
@@ -21,26 +21,38 @@ Representations of functional data are based on three abstract classes. The firs
 Representing Argvals and Values
 ===============================
 
+Functional data representations are based on two main components: the arguments of the functions and the values of the functions. The package provides several classes to represent these components, depending on the type of data to represent. 
+
 .. autosummary::
 	:toctree: autosummary
 
 	FDApy.representation.DenseArgvals
-	FDApy.representation.IrregularArgvals
+	FDApy.representation.DenseValues
+
 
 .. autosummary::
 	:toctree: autosummary
 
-	FDApy.representation.DenseValues
+	FDApy.representation.IrregularArgvals
 	FDApy.representation.IrregularValues
 
 
 Univariate Functional Data
 ==========================
 
-Blablabla
+Univariate functional data are realizations of a random process:
+
+.. math::
+	X: \mathcal{T} \subset \mathbb{R}^d \rightarrow \mathbb{R}.
+
+The package provides two representations of univariate functional data: grid representation and basis representation.
+
 
 Grid representation
 -------------------
+
+Univariate functional data can be represented as a set of values on a grid. The :class:`GridFunctionalData` class is the abstract class to represent univariate functional data on a grid. The package provides two implementations of this class: :class:`DenseFunctionalData` and :class:`IrregularFunctionalData`. The class :class:`DenseFunctionalData` represents functional data of arbitrary dimension (one for curves, two for images, etc.) on a common set of sampling points, while the class :class:`IrregularFunctionalData` represents functional data of arbitrary dimension sampled on different sets of points (the number and location of the sampling points vary between functional observations).
+
 
 .. autosummary::
 	:toctree: autosummary
@@ -53,6 +65,8 @@ Grid representation
 Basis representation
 --------------------
 
+The basis representation of univariate functional data consists of a linear combination of basis functions.
+
 .. autosummary::
 	:toctree: autosummary
 
@@ -62,13 +76,30 @@ Basis representation
 Multivariate Functional Data
 ============================
 
+Multivariate functional data are realizations of a multivariate random process. Multivariate functional data objects are vectors of univariate functional data objects, eventually defined on different domains. The class :class:`MultivariateFunctionalData` allows for the combination of different types of functional data objects (:class:`DenseFunctionalData`, :class:`IrregularFunctionalData`, and :class:`BasisFunctionalData`). It is also possible to mix unidimensional data (curves) with multidimensional data (images, surfaces, etc.).
+
 .. autosummary::
 	:toctree: autosummary
 
[…]
