+++ docs/modules/misc.rst
@@ -2,11 +2,13 @@
 Miscellaneous
 =============
 
-Blablabla
+The package provides several functionalities to facilitate the use of functional data analysis.
 
 
 Visualization
 =============
+
+Basic plots are provided to visualize univariate and multivariate functional data.
 
 .. autosummary::
     :toctree: autosummary
@@ -17,6 +19,8 @@ Visualization
 
 Loader
 ======
+
+The package provides a loader to read CSV files into functional data objects.
 
 .. autosummary::
     :toctree: autosummary
