+++ pyproject.toml
@@ -40,6 +40,7 @@ docs = [
   "scipy",
   "setuptools>=41.2",
   "sphinx>=3",
+  "sphinxcontrib-bibtex",
   "sphinx-gallery",
 ]
 test = [
