### 2025-01-29 StevenGolovkine/FDApy#22 @StevenGolovkine
Hi @vnmabus,

Thank you for your comment on the API reference. I think it greatly improves the package. I'll answer your comment and update my message as long as I'm working on it.

General:

1. I agree with you. I modified the Api reference in this regard.
2. Done.
3. Thank you, I didn't know about that. I'll start doing that while I work on the code.
4. I use the fix of sphinx.ext.napoleon in the conf.py file from skfda. It should be fine now. Thanks.
5. I think I have a problem with inheritance. I'll check everything and make sure the arguments and return values are documented.
6. I'm actually not sure what you call ``vector-valued functions'' and how it differs from multivariate functional data. From the examples in skfda, I think that `VectorValuedBasis` is similar to `MultivariateBasis` in my case.
