### 2025-01-30 StevenGolovkine/FDApy#22 @vnmabus
About 6. Vector-valued functions are functions that can return vectors, that is, functions with the shape $f: \mathcal{T} \to \mathbb{R}^q$ with $q > 1$. Multivariate functional data are just tuples of functions $(f_i: \mathcal{T}_ i \to \mathcal{I}_ i)_{i=1}^M$. When $\mathcal{I}_i = \mathbb{R}$ for every $i$ and $\mathcal{T}_i = \mathcal{T}_j$ for all $i, j$, then there is an isomorphism between the two representations. Thus, having support for multivariate functional data can be enough to support workflows involving vector-valued functions, although having explicit support for the later option  can be more natural in some cases (think for example a function returning a position in 3D space: would you see it more naturally as one vector-valued function, or as three scalar-valued ones?). In any case, I just wanted to know if they were supported or not. I guess the answer is that they are not supported directly, just as multivariate functional data, and that is fine.
