### 2025-01-30 StevenGolovkine/FDApy#22 @StevenGolovkine
Thanks for the explanation! You're right concerning the position in 3D space, it is more natural as one vector-valued functions, like parametric equations. I also think these cases are not really considered in the literature, researchers will use multivariate FD to analyse vector-valued FD (from what I know). Maybe there is something to do in this area.

Concerning the package, for now, I won't include the vector-valued case explicitly and stick with multivariate FD.
