+++ docs/conf.py
@@ -17,6 +17,8 @@ import sys
 from os.path import dirname, relpath
 from typing import Mapping
 
+from sphinx.ext.napoleon.docstring import GoogleDocstring
+
 import FDApy
 #sys.path.insert(0, os.path.abspath(".."))
 
@@ -96,6 +98,11 @@ html_context = {
 
 
 # -- Options for sphinx.ext.autodoc --
+
+
+# -- Options for "sphinx.ext.autodoc.typehints" --
+
+autodoc_typehints = "description"
 
 
 # -- Options for "sphinx.ext.autosummary" --
@@ -187,6 +194,53 @@ def linkcode_resolve(domain: str, info: Mapping[str, str]) -> str | None:
         linespec = ""
 
     return f"{github_url}/tree/{rtd_branch}/FDApy/{fn}{linespec}"
+
+
+# -- Options for "sphinx.ext.napoleon" --
+
+napoleon_use_rtype = True
+
+
+# Napoleon fix for attributes
+# Taken from
+# https://michaelgoerz.net/notes/extending-sphinx-napoleon-docstring-sections.html
+
+
+# first, we define new methods for any new sections and add them to the class
+
+
+def parse_keys_section(self, section):
+    return self._format_fields("Keys", self._consume_fields())
+
+
+GoogleDocstring._parse_keys_section = parse_keys_section
+
+
+def parse_attributes_section(self, section):
+    return self._format_fields("Attributes", self._consume_fields())
+
+
+GoogleDocstring._parse_attributes_section = parse_attributes_section
+
+
+def parse_class_attributes_section(self, section):
+    return self._format_fields("Class Attributes", self._consume_fields())
+
+
+GoogleDocstring._parse_class_attributes_section = parse_class_attributes_section
+
+# we now patch the parse method to guarantee that the the above methods are
+# assigned to the _section dict
+
+
+def patched_parse(self):
+    self._sections["keys"] = self._parse_keys_section
+    self._sections["class attributes"] = self._parse_class_attributes_section
+    self._unpatched_parse()
+
+
+GoogleDocstring._unpatched_parse = GoogleDocstring._parse
+GoogleDocstring._parse = patched_parse
 
 
 # -- Options for "sphinxcontrib.bibtex" --
