### 2025-01-30 #7526 @mstimberg
👋 Hello and a belated Happy New Year to every one. Thanks @vnmabus and @quantgirluk for your review/opened issues so far.
@StevenGolovkine Did you already find the time to look into these issues and/or could you give us an estimate when you will be able to address them? Thanks!
