### 2025-01-30 29364fe
Documentation (Visualization + Misc #22)
