+++ docs/sg_execution_times.rst
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:27.402** total execution time for 28 files **from all galleries**:
+**00:00.000** total execution time for 29 files **from all galleries**:
 
 .. container::
 
@@ -32,87 +32,90 @@ Computation times
    * - Example
      - Time
      - Mem (MB)
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_mfpca_2d.py` (``../examples/fpca/plot_mfpca_2d.py``)
-     - 00:04.282
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_mfpca_1d_2d.py` (``../examples/fpca/plot_mfpca_1d_2d.py``)
-     - 00:04.249
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_mfpca_1d_sparse.py` (``../examples/fpca/plot_mfpca_1d_sparse.py``)
-     - 00:03.496
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_mfpca_1d.py` (``../examples/fpca/plot_mfpca_1d.py``)
-     - 00:03.381
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_smoothing_plot_psplines_2d.py` (``../examples/smoothing/plot_psplines_2d.py``)
-     - 00:02.531
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_fpca_1d_sparse.py` (``../examples/fpca/plot_fpca_1d_sparse.py``)
-     - 00:02.301
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_fpca_2d.py` (``../examples/fpca/plot_fpca_2d.py``)
-     - 00:01.608
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_basis_plot_basis_2d.py` (``../examples/basis/plot_basis_2d.py``)
-     - 00:01.427
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_data_analysis_plot_canadian_weather.py` (``../examples/data_analysis/plot_canadian_weather.py``)
-     - 00:00.722
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_fpca_plot_fpca_1d.py` (``../examples/fpca/plot_fpca_1d.py``)
-     - 00:00.445
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_simulation_plot_karhunen_multivariate.py` (``../examples/simulation/plot_karhunen_multivariate.py``)
-     - 00:00.335
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_smoothing_plot_local_polynomials_2d.py` (``../examples/smoothing/plot_local_polynomials_2d.py``)
-     - 00:00.325
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_simulation_plot_karhunen.py` (``../examples/simulation/plot_karhunen.py``)
-     - 00:00.315
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_representation_plot_multivariate_functional.py` (``../examples/representation/plot_multivariate_functional.py``)
-     - 00:00.301
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_basis_plot_basis_multivariate_1d.py` (``../examples/basis/plot_basis_multivariate_1d.py``)
-     - 00:00.275
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_representation_plot_dense_functional.py` (``../examples/representation/plot_dense_functional.py``)
-     - 00:00.263
-     - 0.0
[…]
