+++ FDApy/misc/utils.py
@@ -65,8 +65,8 @@ DIFF_SEQUENCES = {
 #############################################################################
 def _normalization(
     x: npt.NDArray[np.float64],
-    max_x: Optional[float] = None,
-    min_x: Optional[float] = None,
+    max_x: float | None = None,
+    min_x: float | None = None,
 ) -> npt.NDArray[np.float64]:
     r"""Normalize a vector :math:`[a, b]` into a vector :math:`[0, 1]`.
 
@@ -77,11 +77,11 @@ def _normalization(
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64], shape=(n_obs,)
+    x
         Vector of data
-    max_x: Optional[float], default=None
+    max_x
         Maximum value
-    min_x: Optional[float], default=None
+    min_x
         Minimum value
 
     Returns
@@ -116,7 +116,7 @@ def _standardization(
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64], shape=(n_obs,)
+    x
         Vector of data
 
     Returns
@@ -143,7 +143,7 @@ def _row_mean(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64], shape=(n_obs, n_features)
+    x
         Matrix of data
 
     Returns
@@ -176,7 +176,7 @@ def _row_var(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64], shape=(n_obs, n_features)
+    x
         Matrix of data
 
     Returns
@@ -209,7 +209,7 @@ def _col_mean(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64], shape=(n_obs, n_features)
+    x
         Matrix of data
 
     Returns
@@ -242,7 +242,7 @@ def _col_var(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64], shape=(n_obs, n_features)
+    x
         Matrix of data
 
     Returns
@@ -278,9 +278,9 @@ def _get_axis_dimension(x: npt.NDArray[np.float64], axis: int = 0) -> int:
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64], shape=(n_obs, n_features)
+    x
         Matrix of data
-    axis: int, default=0
+    axis
         Integer value that represents the axis along which the dimension of the
         array is to be computed. The default value is 0.
 
@@ -306,7 +306,7 @@ def _get_dict_dimension(x: Dict[str, npt.NDArray[np.float64]]) -> Tuple[int, ...
 
     Parameters
     ----------
-    x: Dict[str, npt.NDArray[np.float64]]
+    x
         Dictionary containing keys as string and values as numpy array.
 
     Returns
@@ -334,11 +334,11 @@ def _get_obs_shape(
 
     Parameters
     ----------
-    x: Dict[str, Dict[int, npt.NDArray[np.float64]]]
+    x
         Nested dictionary containing the data, where the first level of keys
         are strings and the second level of keys are integers representing the
         observation number.
-    obs: int
+    obs
         Observation number for which to get the shape.
 
     Returns
@@ -371,11 +371,11 @@ def _shift(
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64], shape=(n_obs, n_features)
+    x
         Matrix of data
-    num: int, default=0
+    num
         The number of columns to shift.
-    fill_value: float, default=np.nan
+    fill_value
         The value with one fill the array.
 
     Returns
@@ -413,7 +413,7 @@ def _cartesian_product(
 
[…]
