### 2025-01-30 StevenGolovkine/FDApy#21 @vnmabus
I think this is ok and can be closed.
