+++ FDApy/simulation/brownian.py
@@ -9,7 +9,7 @@ Brownian motions
 import numpy as np
 import numpy.typing as npt
 
-from typing import Callable, Optional, Tuple
+from typing import Callable, Tuple
 
 from ..representation.argvals import DenseArgvals
 from ..representation.values import DenseValues
@@ -31,7 +31,7 @@ def _init_brownian(
 
     Parameters
     ----------
-    argvals: npt.NDArray[np.float64], shape=(n,)
+    argvals
         Values at which Brownian motions are evaluated.
 
     Returns
@@ -54,11 +54,11 @@ def _standard_brownian(
 
     Parameters
     ----------
-    argvals: npt.NDArray[np.float64], shape=(n,)
+    argvals
         Values at which Brownian motions are evaluated.
-    init_point: float, default=0.0
+    init_point
         Start value of the Brownian motion.
-    rnorm: Callable, default=np.random.normal
+    rnorm
         Function to use for normal random variables generation (used for
         reproducibility purpose).
 
@@ -102,16 +102,16 @@ def _geometric_brownian(
 
     Parameters
     ----------
-    argvals: npt.NDArray[np.float64], shape=(n,)
+    argvals
         Values at which Brownian motions are evaluated.
-    init_point: float, default=1.0
+    init_point
         Start value of the Brownian motion. For geometric Brownian motion,
         ``init_point`` should be stricly positive.
-    mu: float, default=0
+    mu
         Interest rate (or percentage drift).
-    sigma: float, default=1
+    sigma
         Diffusion coefficient (or percentage volatility).
-    rnorm: Callable, default=np.random.normal
+    rnorm
         Function to use for normal random variables generation (used for
         reproducibility purpose).
 
@@ -154,12 +154,12 @@ def _fractional_brownian(
 
     Parameters
     ----------
-    argvals: npt.NDArray[np.float64], shape=(n,)
+    argvals
         Values at which Brownian motions are evaluated.
-    hurst: float, default=0.5
+    hurst
         Hurst parameter. If ``hurst = 0.5``. the fractional Brownian motion is
         equivalent to the standard Brownian motion.
-    rnorm: Callable, default=np.random.normal
+    rnorm
         Function to use for normal random variables generation (used for
         reproducibility purpose).
 
@@ -216,14 +216,14 @@ def _simulate_brownian(
 
     Parameters
     ----------
-    name: str, {'standard', 'geometric', 'fractional'}
+    name
         Name of the Brownian motion type to simulate.
-    argvals: npt.NDArray[np.float64], shape=(n,)
+    argvals
         Values at which Brownian motions are evaluated.
-    rnorm: Callable, default=np.random.normal
+    rnorm
         Function to use for normal random variables generation (used for
         reproducibility purpose).
-    **kwargs:
+    kwargs:
         init_point: float, default=0.0 or 1.0
             Start value of the Brownian motion. For geometric Brownian motion,
             ``init_point`` should be stricly positive.
@@ -275,9 +275,9 @@ class Brownian(Simulation):
 
     Parameters
     ----------
-    name: str, {'standard', 'geometric', 'fractional'}
+    name
         Name of the Brownian motion type to simulate.
-    random_state: int, default=None
+    random_state
         A seed to initialize the random number generator.
 
     Attributes
@@ -289,11 +289,13 @@ class Brownian(Simulation):
     sparse_data: IrregularFunctionalData
         An object that represents a sparse version of the simulated data.
 
+    Raises
+    ------
+    ValueError
+        The sampling points have to be regularly spaced. Otherwise, the covariance is not correct.
+
     Notes
     -----
-    The sampling points have to be regularly spaced. Otherwise, the covariance
-    of the generated data will not be the good one.
-
     The implementation is adapted from [1]_.
 
     References
[…]
