+++ FDApy/simulation/karhunen.py
@@ -29,7 +29,7 @@ def _eigenvalues_linear(n: int = 3) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    n: int, default=3
+    n
         Number of eigenvalues to generates.
 
     Returns
@@ -51,7 +51,7 @@ def _eigenvalues_exponential(n: int = 3) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    n: int, default=3
+    n
         Number of eigenvalues to generates.
 
     Returns
@@ -73,7 +73,7 @@ def _eigenvalues_quadratic(n: int = 3) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    n: int, default=3
+    n
         Number of eigenvalues to generates.
 
     Returns
@@ -95,7 +95,7 @@ def _eigenvalues_inverse(n: int = 3) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    n: int, default=3
+    n
         Number of eigenvalues to generates.
 
     Returns
@@ -117,7 +117,7 @@ def _eigenvalues_sqrt(n: int = 3) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    n: int, default=3
+    n
         Number of eigenvalues to generates.
 
     Returns
@@ -139,7 +139,7 @@ def _eigenvalues_wiener(n: int = 3) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    n: int, default=3
+    n
         Number of eigenvalues to generates.
 
     Returns
@@ -161,10 +161,10 @@ def _simulate_eigenvalues(name: str, n: int = 3) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    name: str,
+    name
         Name of the eigenvalues generation process to use. One of
         `{'linear', 'exponential', 'wiener', 'quadratic', 'inverse', 'sqrt'}`.
-    n: int, default=3
+    n
         Number of eigenvalues to generates. Should be strictly positive.
 
     Returns
@@ -209,18 +209,18 @@ def _make_coef(
 
     Parameters
     ----------
-    n_obs: int
+    n_obs
         Number of observations to simulate.
-    n_features: int
+    n_features
         Number of features to simulate.
-    centers: npt.NDArray[np.float64], shape=(n_features, n_clusters)
+    centers
         The centers of the clusters to generate. The ``n_features`` parameter
         corresponds to the number of functions within the basis.
-    clusters_std: npt.NDArray[np.float64], shape=(n_features, n_clusters)
+    clusters_std
         The standard deviation of the clusters to generate. The
         ``n_features`` parameter corresponds to the number of functions within
         the basis.
-    rnorm: Callable, default=np.random.multivariate_normal
+    rnorm
         Method used to generate Gaussian coefficients.
 
     Returns
@@ -268,17 +268,17 @@ def _make_coef(
 
 
 def _initialize_centers(
-    n_features: int, n_clusters: int, centers: Optional[npt.NDArray[np.float64]] = None
+    n_features: int, n_clusters: int, centers: npt.NDArray[np.float64] | None = None
 ) -> npt.NDArray[np.float64]:
     """Initialize the centers of the clusters.
 
     Parameters
     ----------
-    n_features: int
+    n_features
         Number of features to simulate.
-    n_clusters: int
+    n_clusters
         Number of clusters to simulate.
-    centers: Optional[npt.NDArray[np.float64]], shape=(n_features, n_clusters)
+    centers
         The centers of each cluster per feature.
 
     Returns
@@ -294,17 +294,17 @@ def _initialize_centers(
 def _initialize_clusters_std(
     n_features: int,
[…]
