+++ FDApy/preprocessing/smoothing/psplines.py
@@ -18,7 +18,7 @@ from ...misc.basis import _basis_bsplines
 ########################################################################################
 # Utils
 def _row_tensor(
-    x: npt.NDArray[np.float64], y: Optional[npt.NDArray[np.float64]] = None
+    x: npt.NDArray[np.float64], y: npt.NDArray[np.float64] | None = None
 ) -> npt.NDArray[np.float64]:
     """
     Compute the row-wise tensor product of two 2D arrays.
@@ -30,9 +30,9 @@ def _row_tensor(
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64]
+    x
         A 2D array of shape `(m, n)`.
-    y: Optional[npt.NDArray[np.float64]], optional
+    y
         A 2D array of shape `(m, q)`. If not provided, it defaults to `x`.
 
     Returns
@@ -87,9 +87,9 @@ def _h_transform(
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64]
+    x
         A 2D array of shape `(n, m)`.
-    y: npt.NDArray[np.float64]
+    y
         A nD array of shape `(m, n1, n2, ..., nk)`.
 
     Returns
@@ -135,7 +135,7 @@ def _rotate(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64]
+    x
         A multi-dimensional array of shape `(n1, n2, ..., nk)`.
 
     Returns
@@ -179,9 +179,9 @@ def _rotated_h_transform(
 
     Parameters
     ----------
-    x: npt.NDArray[np.float64]
+    x
         A 2D array of shape `(n, m)`.
-    y: npt.NDArray[np.float64]
+    y
         A nD array of shape `(m, n1, n2, ..., nk)`.
 
     Notes
@@ -223,9 +223,9 @@ def _create_permutation(p: int, k: int) -> npt.NDArray[np.float64]:
 
     Parameters
     ----------
-    p: int
+    p
         The number of factors.
-    k: int
+    k
         The number of levels.
 
     Returns
@@ -266,7 +266,7 @@ def _tensor_product_penalties(
 
     Parameters
     ----------
-    penalties: list[npt.NDArray[np.float64]]
+    penalties
         A list of penalty matrices.
 
     Returns
@@ -343,9 +343,9 @@ def _format_data(
 
     Parameters
     ----------
-    X: npt.NDArray[np.float_], shape=(n_obs, n_dimension)
+    X
         An array containing the predictor variable values.
-    y: npt.NDArray[np.float_], shape=(n_obs,)
+    y
         An array containing the response variable values.
 
     """
@@ -369,7 +369,7 @@ def _format_data(
 def _fit_one_dimensional(
     data: npt.NDArray[np.float64],
     basis: npt.NDArray[np.float64],
-    sample_weights: Optional[npt.NDArray[np.float64]] = None,
+    sample_weights: npt.NDArray[np.float64] | None = None,
     penalty: float = 1.0,
     order_penalty: int = 2,
 ) -> Dict[str, npt.NDArray[np.float64]]:
@@ -382,17 +382,17 @@ def _fit_one_dimensional(
 
     Parameters
     ----------
-    data: npt.NDArray[np.float64]
+    data
         A one-dimensional array of shape `(n_obs,)` containing the response variable
         values.
-    basis: npt.NDArray[np.float64]
+    basis
         A two-dimensional array of shape `(n_basis, n_obs)` containing the basis matrix.
-    sample_weights: Optional[npt.NDArray[np.float64]], default=None
+    sample_weights
         A one-dimensional array of shape `(n_obs,)` containing the weights for each
[…]
