+++ FDApy/preprocessing/dim_reduction/mfpca.py
@@ -9,7 +9,7 @@ Multivariate Functional Principal Components Analysis
 import numpy as np
 import numpy.typing as npt
 
-from typing import Dict, Optional, List, Union
+from typing import Dict, List
 
 from ...representation.argvals import DenseArgvals
 from ...representation.values import DenseValues
@@ -39,7 +39,7 @@ from ...misc.utils import _compute_eigen, _block_diag
 def _univariate_decomposition(
     data: FunctionalData,
     method: str = "UFPCA",
-    n_components: Union[int, float] = 2,
+    n_components: int | float = 2,
     **kwargs,
 ) -> BasisFunctionalData:
     """Univariate basis decomposition.
@@ -49,18 +49,17 @@ def _univariate_decomposition(
 
     Parameters
     ----------
-    data: FunctionalData
+    data
         Training data.
-    method: str, default='UFPCA'
+    method
         The method that specify the basis for which the decomposition is to be
         calculated. It should one of `{'UFPCA', 'PSplines', 'FCPTPA'}`.
-    n_components: Union[int, float], default=2
+    n_components
         Number of components to be estimated.
-    **kwargs
-        Other keyword arguments are passed to the following functions:
-
-        - :meth:`UFPCA.fit` and :meth:`UFPCA.transform`;
-        - :meth:`GridFunctionalData.to_basis`.
+    kwargs
+        Other keyword arguments are passed to the functions:
+        :meth:`UFPCA.fit` and :meth:`UFPCA.transform` and
+        :meth:`GridFunctionalData.to_basis`.
 
     Returns
     -------
@@ -114,27 +113,27 @@ def _univariate_decomposition(
 
 def _fit_covariance_multivariate(
     data: MultivariateFunctionalData,
-    n_components: Union[int, float],
+    n_components: int | float,
     univariate_expansions: List[Dict[str, object]],
-    points: Optional[List[DenseArgvals]] = None,
+    points: List[DenseArgvals] | None = None,
 ) -> Dict[str, object]:
     """Multivariate Functional PCA using the covariance operator.
 
     Parameters
     ----------
-    data: FunctionalData
+    data
         Training data.
-    points: DenseArgvals
+    points
         The sampling points at which the covariance and the eigenfunctions
         will be estimated.
-    n_components: Union[int, float]
+    n_components
         Number of components to keep.
-    univariate_expansions: List[Dict[str, object]]
+    univariate_expansions
         List of dictionaries characterizing the univariate expansion computed for each
         component.
-    points: Optional[List[DenseArgvals]]
-            The sampling points at which the covariance and the eigenfunctions
-            will be estimated.
+    points
+        The sampling points at which the covariance and the eigenfunctions
+        will be estimated.
 
     Returns
     -------
@@ -210,24 +209,23 @@ def _fit_covariance_multivariate(
 
 def _fit_inner_product_multivariate(
     data: MultivariateFunctionalData,
-    n_components: Union[int, float] = 1,
-    points: Optional[List[DenseArgvals]] = None,
+    n_components: int | float = 1,
+    points: List[DenseArgvals] | None = None,
     **kwargs,
 ) -> Dict[str, object]:
     """Multivariate Functional PCA using inner-product matrix decomposition.
 
     Parameters
     ----------
-    data: MultivariateFunctionalData
+    data
         Training data used to estimate the eigencomponents.
-    n_components: Union[int, float], default=1
+    n_components
         Number of components to be estimated.
-    points: Optional[List[DenseArgvals]]
+    points
         The sampling points at which the eigenfunctions will be estimated.
-    **kwargs:
-        Other keyword arguments are passed to the following function:
-
-        - :meth:`FunctionalData.inner_product`.
[…]
