+++ FDApy/preprocessing/dim_reduction/fcp_tpa.py
@@ -32,7 +32,7 @@ def _initialize_vectors(
 
     Parameters
     ----------
-    shape: Tuple[int, int, int]
+    shape
         Shape of the dataset. It should be in the format :math:`(N, M_1, M_2)`.
 
     Returns
@@ -53,9 +53,9 @@ def _initalize_output(
 
     Parameters
     ----------
-    shape: Tuple[int, int, int]
+    shape
         Shape of the dataset. It should be in the format :math:`(N, M_1, M_2)`.
-    n_components: int
+    n_components
         Number of components to retain.
 
     Returns
@@ -77,7 +77,7 @@ def _eigendecomposition_penalty_matrices(
 
     Parameters
     ----------
-    penalty_matrices: Dict[str, npt.NDArray[np.float64]]
+    penalty_matrices
         A dictionary with entries :math:`v` and :math:`w`, containing a
         roughness penalty matrix for each direction of the image. The algorithm
         does not induce smoothness along observations.
@@ -109,22 +109,22 @@ def _gcv(
 
     Parameters
     ----------
-    alpha: float
+    alpha
         The current value of the smoothing parameter. It corresponds to
         :math:`\alpha_u` and :math:`\alpha_v` in Equations (19) and (20) in
         [3]_.
-    dimension_length: int
+    dimension_length
         The length of the dimension, for which the smoothing parameter is to
         be optimized. It corresponds to :math:`m` and :math:`n` in Equations
         (19) and (20) in [3]_.
-    vector: npt.NDArray[np.float64], shape=(dimension_length,)
+    vector
         Solutions to the least square problem. It corresponds to
         :math:`Xv / ||v||^2` and :math:`X^\top u / ||u||^2` in Equations (19)
         and (20) in [3]_.
-    smoother: float
+    smoother
         Nornalization parameter. It corresponds to :math:`S_u` and :math:`S_v`
         in Equations (19) and (20) in [3]_.
-    rayleigh: npt.NDArray[np.float64], shape=(dimension_length,)
+    rayleigh
         A vector containing the eigenvalues of the penalty matrix corresponding
         to the current image direction. It corresponds to the Rayleight
         quotients :math:`\mathcal{R}_u(u)` and :math:`\mathcal{R}_v(v)` in
@@ -175,35 +175,35 @@ def _find_optimal_alpha(
 
     Parameters
     ----------
-    alpha_range: Tuple[float, float]
+    alpha_range
         A tuple with two elements, containing the minimal and maximal
         values for the smoothing parameter that is to be optimized. It
         corresponds to minimal and maximal values of :math:`\alpha_u` and
         :math:`\alpha_v` in Equations (19) and (20) in [3]_.
-    data: npt.NDArray[np.float64], shape=(n_obs, m_1, m_2)
+    data
         The tensor containing the data of dimension
         :math:`n_{obs} \times m_1 \times m_2`. It corresponds to
         :math:`\hat{\mathcal{X}}` in Algorithm in [1]_.
-    u: npt.NDArray[np.float64], shape=(n_obs,)
+    u
         The current value of the eigenvectors :math:`u_k` (not
         normalized) of dimensions :math:`n_{obs}`. It corresponds to
         :math:`u_k` in Algorithm in [1]_.
-    v: npt.NDArray[np.float64]
+    v
         The current value of the eigenvectors :math:`v_k` (or :math:`w_k`) (not
         normalized) of dimensions :math:`m_1` (or :math:`m_2`). It corresponds
         to :math:`v_k` (or :math:`w_k`) in Algorithm in [1]_.
-    alpha: float
+    alpha
         The current value of the smoothing parameter for the other image
         direction (:math:`\alpha_w` if the optimization is performed with
[…]
