+++ docs/modules/autosummary/FDApy.representation.DenseFunctionalDataIterator.rst
@@ -1,21 +1,0 @@
-﻿DenseFunctionalDataIterator
-===========================
-
-.. currentmodule:: FDApy.representation
-
-.. autoclass:: DenseFunctionalDataIterator
-
-   
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      
-   
-   
-   
-      
-   
-   
-.. include:: /backreferences/FDApy.representation.DenseFunctionalDataIterator.examples
