+++ docs/modules/autosummary/FDApy.representation.IrregularFunctionalDataIterator.rst
@@ -1,21 +1,0 @@
-﻿IrregularFunctionalDataIterator
-===============================
-
-.. currentmodule:: FDApy.representation
-
-.. autoclass:: IrregularFunctionalDataIterator
-
-   
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      
-   
-   
-   
-      
-   
-   
-.. include:: /backreferences/FDApy.representation.IrregularFunctionalDataIterator.examples

[…]
