+++ FDApy/representation/functional_data.py
@@ -23,11 +23,9 @@ from typing import (
     Callable,
     Dict,
     Iterable,
-    Optional,
     List,
     Tuple,
     Type,
-    Union,
     TYPE_CHECKING,
 )
 
@@ -491,6 +489,17 @@ class GridFunctionalData(FunctionalData):
     values: Type[Values]
         Values of the functional data.
 
+    Attributes
+    ----------
+    argvals_stand: Type[Argvals]
+        Standardized sampling points of the functional data.
+    n_obs: int
+        Number of observations of the functional data.
+    n_dimension: int
+        Number of input dimension of the functional data.
+    n_points: Tuple[int, ...] | Dict[int, Tuple[int, ...]]
+        Number of sampling points.
+
     """
 
     ###########################################################################
@@ -501,7 +510,7 @@ class GridFunctionalData(FunctionalData):
 
         Parameters
         ----------
-        *fdata: FunctionalData
+        fdata
             Functional data to compare.
 
         Raises
@@ -535,7 +544,7 @@ class GridFunctionalData(FunctionalData):
         return (self.argvals == obj.argvals) & np.allclose(self.values, obj.values)
 
     def __add__(
-        self, obj: Union[Type[FunctionalData], float, int]
+        self, obj: Type[FunctionalData] | float | int
     ) -> Type[FunctionalData]:
         """Override add function."""
         if isinstance(obj, FunctionalData):
@@ -546,7 +555,7 @@ class GridFunctionalData(FunctionalData):
             raise TypeError("Operations not available for this type.")
 
     def __sub__(
-        self, obj: Union[Type[FunctionalData], float, int]
+        self, obj: Type[FunctionalData] | float | int
     ) -> Type[FunctionalData]:
         """Override sub function."""
         if isinstance(obj, FunctionalData):
@@ -557,7 +566,7 @@ class GridFunctionalData(FunctionalData):
             raise TypeError("Operations not available for this type.")
 
     def __mul__(
-        self, obj: Union[Type[FunctionalData], float, int]
+        self, obj: Type[FunctionalData] | float | int
     ) -> Type[FunctionalData]:
         """Override mul function."""
         if isinstance(obj, FunctionalData):
@@ -568,13 +577,13 @@ class GridFunctionalData(FunctionalData):
             raise TypeError("Operations not available for this type.")
 
     def __rmul__(
-        self, obj: Union[Type[FunctionalData], float, int]
+        self, obj: Type[FunctionalData] | float | int
     ) -> Type[FunctionalData]:
         """Override rmul function."""
         return self * obj
 
     def __truediv__(
-        self, obj: Union[Type[FunctionalData], float, int]
+        self, obj: Type[FunctionalData] | float | int
     ) -> Type[FunctionalData]:
         """Override truediv function."""
         if isinstance(obj, FunctionalData):
@@ -585,7 +594,7 @@ class GridFunctionalData(FunctionalData):
             raise TypeError("Operations not available for this type.")
 
     def __floordiv__(
-        self, obj: Union[Type[FunctionalData], float, int]
+        self, obj: Type[FunctionalData] | float | int
     ) -> Type[FunctionalData]:
         """Override floordiv function."""
         if isinstance(obj, FunctionalData):
@@ -656,12 +665,12 @@ class GridFunctionalData(FunctionalData):
         return self.argvals.n_dimension
 
     @property
-    def n_points(self) -> Union[Tuple[int, ...], Dict[int, Tuple[int, ...]]]:
+    def n_points(self) -> Tuple[int, ...] | Dict[int, Tuple[int, ...]]:
         """Get the number of sampling points.
 
         Returns
         -------
[…]
