+++ FDApy/representation/functional_data.py
@@ -230,7 +230,7 @@ def _estimate_noise_variance_with_covariance(
 # Class FunctionalData
 class FunctionalData(ABC):
     """Define the structure of FunctionalData.
-    
+
     Attributes
     ----------
     n_obs: int
@@ -543,9 +543,7 @@ class GridFunctionalData(FunctionalData):
             raise TypeError("Object does not have the right type.")
         return (self.argvals == obj.argvals) & np.allclose(self.values, obj.values)
 
-    def __add__(
-        self, obj: Type[FunctionalData] | float | int
-    ) -> Type[FunctionalData]:
+    def __add__(self, obj: Type[FunctionalData] | float | int) -> Type[FunctionalData]:
         """Override add function."""
         if isinstance(obj, FunctionalData):
             return self._perform_computation(self, obj, np.add)
@@ -554,9 +552,7 @@ class GridFunctionalData(FunctionalData):
         else:
             raise TypeError("Operations not available for this type.")
 
-    def __sub__(
-        self, obj: Type[FunctionalData] | float | int
-    ) -> Type[FunctionalData]:
+    def __sub__(self, obj: Type[FunctionalData] | float | int) -> Type[FunctionalData]:
         """Override sub function."""
         if isinstance(obj, FunctionalData):
             return self._perform_computation(self, obj, np.subtract)
@@ -565,9 +561,7 @@ class GridFunctionalData(FunctionalData):
         else:
             raise TypeError("Operations not available for this type.")
 
-    def __mul__(
-        self, obj: Type[FunctionalData] | float | int
-    ) -> Type[FunctionalData]:
+    def __mul__(self, obj: Type[FunctionalData] | float | int) -> Type[FunctionalData]:
         """Override mul function."""
         if isinstance(obj, FunctionalData):
             return self._perform_computation(self, obj, np.multiply)
@@ -576,9 +570,7 @@ class GridFunctionalData(FunctionalData):
         else:
             raise TypeError("Operations not available for this type.")
 
-    def __rmul__(
-        self, obj: Type[FunctionalData] | float | int
-    ) -> Type[FunctionalData]:
+    def __rmul__(self, obj: Type[FunctionalData] | float | int) -> Type[FunctionalData]:
         """Override rmul function."""
         return self * obj
 
@@ -955,7 +947,7 @@ class DenseFunctionalData(GridFunctionalData):
         ----------
         fdata
             Functional data to concatenate.
-        
+
         Returns
         -------
         DenseFunctionalData
@@ -1295,7 +1287,7 @@ class DenseFunctionalData(GridFunctionalData):
             performed. If 'PS', the method is P-splines [4]_. If 'LP', the
             method is local polynomials [8]_.
         kwargs
-            Other keyword arguments are passed to the following function 
+            Other keyword arguments are passed to the following function
             :meth:`DenseFunctionalData.smooth`.
 
         Returns
@@ -3103,7 +3095,7 @@ class BasisFunctionalData(FunctionalData):
     # Methods
     def to_grid(self) -> DenseFunctionalData:
         """Convert the data to grid format.
-        
+
         Returns
         -------
         DenseFunctionalData
@@ -3116,7 +3108,7 @@ class BasisFunctionalData(FunctionalData):
 
     def to_long(self, reindex: bool = False) -> pd.DataFrame:
         """Convert the data to long format.
-        
+
         Parameters
         ----------
         reindex
@@ -3211,7 +3203,7 @@ class BasisFunctionalData(FunctionalData):
         **kwargs,
     ) -> FunctionalData:
         """Compute an estimate of the mean.
-        
+
         Parameters
         ----------
         points
@@ -3222,14 +3214,14 @@ class BasisFunctionalData(FunctionalData):
[…]
