+++ FDApy/representation/basis.py
@@ -15,7 +15,7 @@ import numpy.typing as npt
 from functools import reduce
 from scipy.integrate import simpson
 
-from typing import Optional, List, Tuple, Union
+from typing import List, Tuple
 
 from .functional_data import DenseFunctionalData, MultivariateFunctionalData
 
@@ -40,20 +40,19 @@ def _simulate_basis(
 
     Parameters
     ----------
-    name: str, {'legendre', 'wiener', 'fourier', 'bsplines'}
+    name
         Name of the basis to use.
-    argvals: npt.NDArray[np.float64]
+    argvals
         The values on which the basis functions are evaluated.
-    n_functions: int, default=5
+    n_functions
         Number of functions to compute.
-    is_normalized: bool
+    is_normalized
         Should we normalize the functions?
-    add_intercept: bool, default=True
+    add_intercept
         Should the constant functions be into the basis?
     kwargs
-        Other keyword arguments are passed to the function:
-
-        - :meth:`misc.basis._basis_bsplines`.
+        Other keyword arguments are passed to the function
+        :meth:`misc.basis._basis_bsplines`.
 
     Returns
     -------
@@ -107,25 +106,35 @@ class Basis(DenseFunctionalData):
 
     Parameters
     ----------
-    name: Union[Tuple[str], str], {'given', 'legendre', 'wiener', 'fourier', 'bsplines'}
+    name
         Denotes the basis of functions to use. The default is `bsplines`. If
         `name=given`, it uses a user defined basis (defined with the `argvals` and
         `values` parameters). For higher dimensional data, `name` is a tuple for the
         marginal basis.
-    n_functions: Union[Tuple[int], int], default=5
+    n_functions
         Number of functions in the basis.
-    argvals: Optional[DenseArgvals]
+    argvals
         The sampling points of the functional data.
-    values: Optional[DenseValues]
+    values
         The values of the functional data. Only used if `name='given'`.
-    is_normalized: bool, default=False
+    is_normalized
         Should we normalize the basis function?
-    add_intercept: bool, default=True
+    add_intercept
         Should the constant functions be into the basis?
     kwargs
-        Other keyword arguments are passed to the function:
+        Other keyword arguments are passed to the function
+        :meth:`representation.basis._simulate_basis`.
 
-        - :meth:`representation.basis._simulate_basis`.
+    Attributes
+    ----------
+    argvals_stand: DenseArgvals
+        Standardized sampling points of the functional data.
+    n_obs: int
+        Number of observations of the functional data.
+    n_dimension: int
+        Number of input dimension of the functional data.
+    n_points: Tuple[int, ...]
+        Number of sampling points.
 
     References
     ----------
@@ -156,10 +165,10 @@ class Basis(DenseFunctionalData):
     # Magic methods
     def __init__(
         self,
-        name: Union[Tuple[str], str] = "bsplines",
-        n_functions: Union[Tuple[int], int] = 5,
-        argvals: Optional[DenseArgvals] = None,
-        values: Optional[DenseValues] = None,
+        name: Tuple[str] | str = "bsplines",
+        n_functions: Tuple[int] | int = 5,
+        argvals: DenseArgvals | None = None,
+        values: DenseValues | None = None,
         is_normalized: bool = False,
         add_intercept: bool = True,
         **kwargs,
@@ -207,7 +216,7 @@ class Basis(DenseFunctionalData):
         return self._name
 
     @name.setter
-    def name(self, new_name: Union[Tuple[str], str]) -> None:
+    def name(self, new_name: Tuple[str] | str) -> None:
         if isinstance(new_name, str):
             new_name = (new_name,)
         self._name = new_name
[…]
