### 2025-02-01 StevenGolovkine/FDApy#22 @StevenGolovkine
For the other comments:

Representation:
1. The name `n_dimension` refers to the input dimension. I added an attribute in the documentation of the functional data classes to make it clear.
2. I will continue to use ``observation'' for a realization of a functional random variable and use ``sampling points'' for the observed points within an observation of a functional data.
3. Fixed.
4. A format to represent functional data as pd.Dataframe. In some cases, it is easier to manipulate this format for calculation. I added some explanation in the doc and will add an example to explain it.
5. The noise variance is estimated using the difference between values at consecutive sampling points. The `order` is the number of points to consider (see Hall, P., Kay, J.W. and Titterington, D.M. (1990). Asymptotically Optimal Difference-Based Estimation of Variance in Nonparametric Regression. Biometrika 77, 521–528.)
6. While it is mostly interesting for IrregularFunctionalData, some smoothing is also include in the estimation of the mean (and to the centering) for DenseFunctionalData (see Cai, T.T., Yuan, M., (2011), Optimal estimation of the mean function based on discretely sampled functional data: Phase transition. The Annals of Statistics 39, 2330-2355.)
7. I plan to include other norms at some points. The formula of the norm is written in the doc.
8. I would say it is not really useful for univariate functional data. However, for multivariate functional data, it would make sense to standardize each individual feature if they have very different scale, even to run MFPCA, so the resulting functional eigencomponents won't represent only the features with the most variance.
9. Rescaling and standardization are similar. The only difference is how the weights are computed. For standardization, the weights are :math:`C(t, t)^{-½}`, while for the rescaling, the weights can be user defined.
10. A version of the argvals in [0, 1]. I've had documentation of for the DenseArgvals and IrregularArgvals.
11. Thanks, fixed.
12. The default bandwidth is the optimal value if the curves are sampled on [0, 1]. I don't really have an answer on why not forcing the user to choose a bandwidth, it is just a choice I've made a long time ago.
13. Fixed.
13. Fixed.
14. Fixed
15. I'll add an example with the different available basis.

Preprocessing:
1. Ok

Visualization:
1. Fixed.
2. Yes, it is. I changed the doc.

Simulation:
1. Ok, it now raises an exception.

Miscellaneous:
1. I'll write the format of the CSV in the examples. Add comment on the returning format.
