+++ FDApy/misc/loader.py
@@ -26,7 +26,10 @@ def read_csv(filepath: str, **kwargs) -> DenseFunctionalData | IrregularFunction
     Build a DenseFunctionalData or IrregularFunctionalData object upon a CSV
     file passed as parameter. If the CSV file does not contain any `NA` values, the
     data will be loaded as a DenseFunctionalData object. Otherwise, it will be loaded
-    as an IrregularFunctionalData object. See the `Canadian Weather example <../../auto_examples/data_analysis/plot_canadian_weather.html>`_ and `CD4 example <../../auto_examples/data_analysis/plot_cd4.html>`_ for the formating of the CSV file.
+    as an IrregularFunctionalData object. See the `Canadian Weather example
+    <../../auto_examples/data_analysis/plot_canadian_weather.html>`_ and
+    `CD4 example <../../auto_examples/data_analysis/plot_cd4.html>`_ for the formating
+    of the CSV file.
 
 
     Notes
