+++ docs/modules/autosummary/FDApy.representation.Argvals.rst
@@ -1,105 +1,0 @@
-﻿Argvals
-=======
-
-.. currentmodule:: FDApy.representation
-
-.. autoclass:: Argvals
-
-   
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      
-      
-      ~Argvals.clear
-      
-      
-      ~Argvals.compatible_with
-      
-      
-      ~Argvals.concatenate
-      
-      
-      ~Argvals.copy
-      
-      
-      ~Argvals.fromkeys
-      
-      
-      ~Argvals.get
-      
-      
-      ~Argvals.items
-      
-      
-      ~Argvals.keys
-      
-      
-      ~Argvals.normalization
-      
-      
-      ~Argvals.pop
-      
-      
-      ~Argvals.popitem
-      
-      
-      ~Argvals.setdefault
-      
-      
-      ~Argvals.update
-      
-      
-      ~Argvals.values
-      
-   
-   
-   
-      
-      
-   .. automethod:: clear
-      
-      
-   .. automethod:: compatible_with
-      
-      
-   .. automethod:: concatenate
-      
-      
-   .. automethod:: copy
-      
-      
-   .. automethod:: fromkeys
-      
-      
-   .. automethod:: get
-      
-      
-   .. automethod:: items
-      
-      
-   .. automethod:: keys
-      
-      
-   .. automethod:: normalization
-      
-      
-   .. automethod:: pop
-      
-      
-   .. automethod:: popitem
-      
-      
-   .. automethod:: setdefault
-      
-      
-   .. automethod:: update
-      
-      
-   .. automethod:: values
-      
-   
-   
-.. include:: /backreferences/FDApy.representation.Argvals.examples
