+++ docs/modules/autosummary/FDApy.representation.FunctionalData.rst
@@ -1,93 +1,0 @@
-﻿FunctionalData
-==============
-
-.. currentmodule:: FDApy.representation
-
-.. autoclass:: FunctionalData
-
-   
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      
-      
-      ~FunctionalData.center
-      
-      
-      ~FunctionalData.concatenate
-      
-      
-      ~FunctionalData.covariance
-      
-      
-      ~FunctionalData.inner_product
-      
-      
-      ~FunctionalData.mean
-      
-      
-      ~FunctionalData.noise_variance
-      
-      
-      ~FunctionalData.norm
-      
-      
-      ~FunctionalData.normalize
-      
-      
-      ~FunctionalData.rescale
-      
-      
-      ~FunctionalData.smooth
-      
-      
-      ~FunctionalData.standardize
-      
-      
-      ~FunctionalData.to_long
-      
-   
-   
-   
-      
-      
-   .. automethod:: center
-      
-      
-   .. automethod:: concatenate
-      
-      
-   .. automethod:: covariance
-      
-      
-   .. automethod:: inner_product
-      
-      
-   .. automethod:: mean
-      
-      
-   .. automethod:: noise_variance
-      
-      
-   .. automethod:: norm
-      
-      
-   .. automethod:: normalize
-      
-      
-   .. automethod:: rescale
-      
-      
-   .. automethod:: smooth
-      
-      
-   .. automethod:: standardize
-      
-      
-   .. automethod:: to_long
-      
-   
-   
-.. include:: /backreferences/FDApy.representation.FunctionalData.examples
