+++ docs/modules/autosummary/FDApy.misc.read_csv.rst
@@ -1,8 +1,0 @@
-﻿read\_csv
-=========
-
-.. currentmodule:: FDApy.misc
-
-.. autofunction:: read_csv
-
-.. include:: /backreferences/FDApy.misc.read_csv.examples
