+++ docs/modules/autosummary/FDApy.representation.DenseArgvals.rst
@@ -1,111 +1,0 @@
-﻿DenseArgvals
-============
-
-.. currentmodule:: FDApy.representation
-
-.. autoclass:: DenseArgvals
-
-   
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      
-      
-      ~DenseArgvals.clear
-      
-      
-      ~DenseArgvals.compatible_with
-      
-      
-      ~DenseArgvals.concatenate
-      
-      
-      ~DenseArgvals.copy
-      
-      
-      ~DenseArgvals.fromkeys
-      
-      
-      ~DenseArgvals.get
-      
-      
-      ~DenseArgvals.items
-      
-      
-      ~DenseArgvals.keys
-      
-      
-      ~DenseArgvals.normalization
-      
-      
-      ~DenseArgvals.pop
-      
-      
-      ~DenseArgvals.popitem
-      
-      
-      ~DenseArgvals.range
-      
-      
-      ~DenseArgvals.setdefault
-      
-      
-      ~DenseArgvals.update
-      
-      
-      ~DenseArgvals.values
-      
-   
-   
-   
-      
-      
-   .. automethod:: clear
-      
-      
-   .. automethod:: compatible_with
-      
-      
-   .. automethod:: concatenate
-      
-      
-   .. automethod:: copy
-      
-      
-   .. automethod:: fromkeys
-      
-      
-   .. automethod:: get
-      
-      
-   .. automethod:: items
-      
-      
-   .. automethod:: keys
-      
-      
-   .. automethod:: normalization
-      
-      
-   .. automethod:: pop
-      
-      
-   .. automethod:: popitem
-      
-      
-   .. automethod:: range
-      
-      
-   .. automethod:: setdefault
-      
-      
-   .. automethod:: update
-      
-      
-   .. automethod:: values
-      
-   
-   
-.. include:: /backreferences/FDApy.representation.DenseArgvals.examples
