+++ docs/modules/autosummary/FDApy.visualization.plot_multivariate.rst
@@ -1,8 +1,0 @@
-﻿plot\_multivariate
-==================
-
-.. currentmodule:: FDApy.visualization
-
-.. autofunction:: plot_multivariate
-
-.. include:: /backreferences/FDApy.visualization.plot_multivariate.examples
