+++ docs/modules/autosummary/FDApy.simulation.Datasets.rst
@@ -1,45 +1,0 @@
-﻿Datasets
-========
-
-.. currentmodule:: FDApy.simulation
-
-.. autoclass:: Datasets
-
-   
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      
-      
-      ~Datasets.add_noise
-      
-      
-      ~Datasets.add_noise_and_sparsify
-      
-      
-      ~Datasets.new
-      
-      
-      ~Datasets.sparsify
-      
-   
-   
-   
-      
-      
-   .. automethod:: add_noise
-      
-      
-   .. automethod:: add_noise_and_sparsify
-      
-      
-   .. automethod:: new
-      
-      
-   .. automethod:: sparsify
-      
-   
-   
-.. include:: /backreferences/FDApy.simulation.Datasets.examples

[…]
