### 2025-02-01 01292cc
Type hints
