### 2025-02-01 5e82941
Add packages version (#27)
