+++ CHANGELOG.rst
@@ -1,0 +1,14 @@
+=========
+Changelog
+=========
+
+Version 1.0.2 (2024-09-24)
+==========================
+
+- Add P-splines smoothing method.
+- Add basis representation for functional data.
+
+Version 1.0.0 (2023-12-29)
+==========================
+
+- Initial release.
