### 2025-02-01 StevenGolovkine/FDApy#27 @StevenGolovkine
Hi @quantgirluk,

Thanks for your comments. I've changed the documentation regarding the other reviewer comments (see #20, #21, #22).
I've added package requirements in the documentation main page and added a CHANGELOG.
The documentation is for v1.0.2, I forgot to update the version. The added functionalities between v1.0.0 and v1.0.2 is in the change log.

Let me know if it requires other changes.
