+++ setup.cfg
@@ -6,7 +6,6 @@ universal = 1
 strict = True
 strict_equality = True
 implicit_reexport = True
-plugins = numpy.typing.mypy_plugin
 
 [mypy-pandas.*]
 ignore_missing_imports = True
@@ -15,6 +14,9 @@ ignore_missing_imports = True
 ignore_missing_imports = True
 
 [mypy-sklearn.*]
+ignore_missing_imports = True
+
+[mypy-lazy_loader.*]
 ignore_missing_imports = True
 
 [flake8]
