+++ FDApy/misc/utils.py
@@ -10,7 +10,7 @@ import numpy as np
 import numpy.typing as npt
 import scipy
 
-from typing import Callable, Dict, List, Tuple
+from typing import Callable, Dict, List, Literal, Tuple
 
 #############################################################################
 # Constants
@@ -95,9 +95,10 @@ def _normalization(
     array([0., 0.5, 1.])
 
     """
-    if (max_x is None) and (min_x is None):
-        max_x = np.amax(x)
-        min_x = np.amin(x)
+    if max_x is None:
+        max_x = max(x)
+    if min_x is None:
+        min_x = min(x)
     if max_x == min_x:
         return np.zeros_like(x)
     else:
@@ -133,7 +134,8 @@ def _standardization(
     if np.std(x) == 0:
         return np.zeros_like(x)
     else:
-        return (x - np.mean(x)) / np.std(x)
+        result: npt.NDArray[np.float64] = (x - np.mean(x)) / np.std(x)
+        return result
 
 
 def _row_mean(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
@@ -166,7 +168,8 @@ def _row_mean(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
     array([1., 2., 3.])
 
     """
-    return np.mean(x, axis=0)
+    result: npt.NDArray[np.float64] = np.mean(x, axis=0)
+    return result
 
 
 def _row_var(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
@@ -199,7 +202,8 @@ def _row_var(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
     array([0., 0., 0.])
 
     """
-    return x.var(axis=0)
+    result: npt.NDArray[np.float64] = np.var(x, axis=0)
+    return result
 
 
 def _col_mean(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
@@ -232,7 +236,8 @@ def _col_mean(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
     array([2., 2., 2., 2.])
 
     """
-    return x.mean(axis=1)
+    result: npt.NDArray[np.float64] = np.mean(x, axis=1)
+    return result
 
 
 def _col_var(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
@@ -265,7 +270,8 @@ def _col_var(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
     array([0.66666667, 0.66666667, 0.66666667, 0.66666667])
 
     """
-    return x.var(axis=1)
+    result: npt.NDArray[np.float64] = np.var(x, axis=1)
+    return result
 
 
 ############################################################################
@@ -459,7 +465,7 @@ def _cartesian_product(
     return stacked
 
 
-def _block_diag(*arrs: npt.NDArray[np.float_]):
+def _block_diag(*arrs: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
     """Create a block diagonal matrix from provided arrays.
 
     Given the inputs `A`, `B` and `C`, the output will have these
@@ -532,7 +538,8 @@ def _block_diag(*arrs: npt.NDArray[np.float_]):
 
 
 def _integration_weights(
-    x: npt.NDArray[np.float64], method: str | Callable = "trapz"
+    x: npt.NDArray[np.float64],
+    method: str | Callable[[npt.ArrayLike], npt.ArrayLike] = "trapz"
 ) -> npt.NDArray[np.float64]:
     """Compute integration weights.
 
@@ -640,7 +647,7 @@ def _integrate(
     else:
         raise ValueError(f"{method} not implemented!")
 
[…]
