+++ setup.cfg
@@ -6,6 +6,7 @@ universal = 1
 strict = True
 strict_equality = True
 implicit_reexport = True
+plugins = numpy.typing.mypy_plugin
 
 [mypy-pandas.*]
 ignore_missing_imports = True
