### 2025-02-02 c874437
Typing preprocessing/smoothing/local_polynomial.py
