+++ FDApy/preprocessing/smoothing/local_polynomial.py
@@ -6,8 +6,6 @@ Local Polynomials
 -----------------
 
 """
-from __future__ import annotations
-
 import numpy as np
 import numpy.typing as npt
 
@@ -16,11 +14,13 @@ from typing import Callable
 from sklearn.preprocessing import PolynomialFeatures
 
 
+NDArrayFloat = npt.NDArray[np.float64]
+
 ##############################################################################
 # Inner functions for the LocalPolynomial class.
 
 
-def _gaussian(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
+def _gaussian(x: NDArrayFloat) -> NDArrayFloat:
     r"""Compute the Gaussian density with mean 0 and standard deviation 1.
 
     The Gaussian density is given by
@@ -35,14 +35,14 @@ def _gaussian(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Returns
     -------
-    npt.NDArray[np.float64], shape = (n_samples,)
+    NDArrayFloat, shape = (n_samples,)
         Values of the kernel.
 
     """
-    return np.exp(-np.square(x) / 2) / np.sqrt(2 * np.pi)
+    return np.exp(-np.square(x) / 2) / np.sqrt(2 * np.pi)  # type: ignore
 
 
-def _epanechnikov(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
+def _epanechnikov(x: NDArrayFloat) -> NDArrayFloat:
     r"""Compute the Epanechnikov kernel.
 
     The Epanechnikov kernel is given, in [1]_ equation 6.4, by
@@ -57,7 +57,7 @@ def _epanechnikov(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Returns
     -------
-    npt.NDArray[np.float64], shape = (n_samples,)
+    NDArrayFloat, shape = (n_samples,)
         Values of the kernel.
 
     References
@@ -73,7 +73,7 @@ def _epanechnikov(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
     return kernel
 
 
-def _tri_cube(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
+def _tri_cube(x: NDArrayFloat) -> NDArrayFloat:
     r"""Compute the tri-cube kernel.
 
     The tri-cube kernel is given, in [1]_ equation 6.6, by
@@ -88,7 +88,7 @@ def _tri_cube(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Returns
     -------
-    npt.NDArray[np.float64], shape = (n_samples,)
+    NDArrayFloat, shape = (n_samples,)
         Values of the kernel.
 
     References
@@ -104,7 +104,7 @@ def _tri_cube(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
     return kernel
 
 
-def _bi_square(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
+def _bi_square(x: NDArrayFloat) -> NDArrayFloat:
     r"""Compute the bi-square kernel.
 
     The bi-square kernel is given, in [1]_, by
@@ -119,7 +119,7 @@ def _bi_square(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Returns
     -------
-    npt.NDArray[np.float64], shape = (n_samples,)
+    NDArrayFloat, shape = (n_samples,)
         Values of the kernel.
 
     References
@@ -135,7 +135,7 @@ def _bi_square(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
     return kernel
[…]
