### 2025-02-03 21dc009
Typing preprocessing/smoothing/psplines.py
