+++ FDApy/preprocessing/smoothing/psplines.py
@@ -8,18 +8,23 @@ P-splines
 """
 import numpy as np
 import numpy.typing as npt
+import sys
 
-
-from typing import Dict, List, Union
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    from typing_extensions import Self
 
 from ...misc.basis import _basis_bsplines
 
 
+NDArrayFloat = npt.NDArray[np.float64]
+
 ########################################################################################
 # Utils
 def _row_tensor(
-    x: npt.NDArray[np.float64], y: npt.NDArray[np.float64] | None = None
-) -> npt.NDArray[np.float64]:
+    x: NDArrayFloat, y: NDArrayFloat | None = None
+) -> NDArrayFloat:
     """
     Compute the row-wise tensor product of two 2D arrays.
 
@@ -37,7 +42,7 @@ def _row_tensor(
 
     Returns
     -------
-    npt.NDArray[np.float64]
+    NDArrayFloat
         A 2D array of shape `(m, n*q)` or `(m, n*n)` if `y` is not provided.
 
     Examples
@@ -76,8 +81,8 @@ def _row_tensor(
 
 
 def _h_transform(
-    x: npt.NDArray[np.float64], y: npt.NDArray[np.float64]
-) -> npt.NDArray[np.float64]:
+    x: NDArrayFloat, y: NDArrayFloat
+) -> NDArrayFloat:
     """
     Compute the H-transform of a nD array `y` with respect to a 2D array `x`.
 
@@ -94,7 +99,7 @@ def _h_transform(
 
     Returns
     -------
-    npt.NDArray[np.float64]
+    NDArrayFloat
         A nD array of shape `(n, n1, n2, ..., nk)`.
 
     Notes
@@ -125,7 +130,7 @@ def _h_transform(
     return xy_product.reshape((xy_product.shape[0], *y_dim[1:]))
 
 
-def _rotate(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
+def _rotate(x: NDArrayFloat) -> NDArrayFloat:
     """
     Rotate the axes of a multi-dimensional array to the right.
 
@@ -140,7 +145,7 @@ def _rotate(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
     Returns
     -------
-    npt.NDArray[np.float64]
+    NDArrayFloat
         A multi-dimensional array of shape `(n2, ..., nk, n1)`.
 
     Notes
@@ -168,8 +173,8 @@ def _rotate(x: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
 
 
 def _rotated_h_transform(
-    x: npt.NDArray[np.float64], y: npt.NDArray[np.float64]
-) -> npt.NDArray[np.float64]:
+    x: NDArrayFloat, y: NDArrayFloat
+) -> NDArrayFloat:
     """
     Compute the rotated H-transform of a nD array `y` with respect to a 2D array `x`.
 
@@ -190,7 +195,7 @@ def _rotated_h_transform(
 
     Returns
     -------
-    npt.NDArray[np.float64]
+    NDArrayFloat
         A nD array of shape `(n1, n2, ..., nk, m)`.
 
     Examples
@@ -213,7 +218,7 @@ def _rotated_h_transform(
     return _rotate(_h_transform(x, y))
 
 
-def _create_permutation(p: int, k: int) -> npt.NDArray[np.float64]:
+def _create_permutation(p: int, k: int) -> NDArrayFloat:
     """
     Create a permutation array for a given number of factors and levels.
 
@@ -230,7 +235,7 @@ def _create_permutation(p: int, k: int) -> npt.NDArray[np.float64]:
 
     Returns
     -------
[…]
