+++ pyproject.toml
@@ -23,7 +23,8 @@ dependencies = [
   'numpy<2.0.0',
   'pandas>=2.0.0',
   'scikit-learn>=1.2.0',
-  'scipy>=1.10.0'
+  'scipy>=1.10.0',
+  'typing_extensions; python_version<3.11'
 ]
 
 [project.optional-dependencies]
