+++ pyproject.toml
@@ -24,7 +24,7 @@ dependencies = [
   'pandas>=2.0.0',
   'scikit-learn>=1.2.0',
   'scipy>=1.10.0',
-  'typing_extensions; python_version<3.11'
+  'typing_extensions; python_version<"3.11"'
 ]
 
 [project.optional-dependencies]
