### 2025-02-03 8e77221
Typing representation/values.py
