+++ FDApy/representation/argvals.py
@@ -13,32 +13,30 @@ import numpy.typing as npt
 
 from abc import abstractmethod
 from collections import UserDict
-from typing import Any, Dict, Tuple, TYPE_CHECKING
+from typing import Any, TYPE_CHECKING
 
 if TYPE_CHECKING:
     from .values import Values
 
 
+NDArrayFloat = npt.NDArray[np.float64]
+
+
 ###############################################################################
 # Class Argvals
-class Argvals(UserDict):
+class Argvals(UserDict[Any, Any]):
     """Define the structure of Argvals.
 
     Attributes
     ----------
-    n_points: Tuple[int, ...] | Dict[int, Tuple[int, ...]]
+    n_points: tuple[int, ...] | dict[int, tuple[int, ...]]
         Number of sampling points of each dimension.
     n_dimension: int
         Number of input dimension of the data.
-    min_max: Dict[str, Tuple[float, float]]
+    min_max: dict[str, tuple[float, float]]
         Minimum and maximum sampling points for each dimension.
 
     """
-
-    @staticmethod
-    @abstractmethod
-    def concatenate(*argvals: Argvals) -> Argvals:
-        """Concatenate Argvals objects."""
 
     @abstractmethod
     def __setitem__(self, key: Any, value: Any) -> None:
@@ -63,7 +61,7 @@ class Argvals(UserDict):
         super().__setitem__(key, value)
 
     @abstractmethod
-    def __eq__(self, other: Argvals) -> bool:
+    def __eq__(self, other: Any) -> bool:
         """Check if two Argvals are equals.
 
         This method if two Argvals objects have the same type and if their
@@ -89,21 +87,21 @@ class Argvals(UserDict):
 
     @property
     @abstractmethod
-    def n_points(self):
+    def n_points(self) -> tuple[int, ...] | dict[int, tuple[int, ...]]:
         """Get the number of sampling points of each dimension."""
 
     @property
     @abstractmethod
-    def n_dimension(self):
+    def n_dimension(self) -> int:
         """Get the number of dimension of the data."""
 
     @property
     @abstractmethod
-    def min_max(self):
+    def min_max(self) -> dict[str, tuple[float, float]]:
         """Get the minimum and maximum sampling points for each dimension."""
 
     @abstractmethod
-    def normalization(self):
+    def normalization(self) -> Argvals:
         """Normalize the Argvals."""
 
     def compatible_with(self, values: Values) -> None:
@@ -140,17 +138,17 @@ class DenseArgvals(Argvals):
 
     Attributes
     ----------
-    n_points: Tuple[int, ...]
+    n_points: tuple[int, ...]
         Number of sampling points of each dimension.
     n_dimension: int
         Number of input dimension of the data.
-    min_max: Dict[str, Tuple[float, float]]
+    min_max: dict[str, tuple[float, float]]
         Minimum and maximum sampling points for each dimension.
 
     """
 
     @staticmethod
-    def concatenate(*argvals) -> DenseArgvals:
+    def concatenate(*argvals: DenseArgvals) -> DenseArgvals:
         """Concatenate DenseArgvals objects.
 
         It does not make sense to concatenate DenseArgvals. This function
@@ -208,7 +206,7 @@ class DenseArgvals(Argvals):
             raise TypeError("Value must be an np.ndarray")
         super().__setitem__(key, value)
 
-    def __eq__(self, other: DenseArgvals) -> bool:
+    def __eq__(self, other: Any) -> bool:
         """Check if two DenseArgvals objects are equal.
 
         This method compares the DenseArgvals object with another object to
@@ -255,7 +253,7 @@ class DenseArgvals(Argvals):
         return True
 
     @property
[…]
