+++ FDApy/simulation/karhunen.py
@@ -492,7 +492,7 @@ class KarhunenLoeve(Simulation):
         n_obs: int,
         n_clusters: int = 1,
         argvals: npt.NDArray[np.float64] | None = None,
-        **kwargs,
+        **kwargs: Any,
     ) -> None:
         """Simulate realizations from Karhunen-Loève decomposition.
 
