+++ FDApy/representation/functional_data.py
@@ -19,15 +19,7 @@ from functools import reduce
 from abc import ABC, abstractmethod
 from collections import UserList
 from collections.abc import Iterator
-from typing import (
-    Callable,
-    Dict,
-    Iterable,
-    List,
-    Tuple,
-    Type,
-    TYPE_CHECKING,
-)
+from typing import Any, Callable, Iterable, Type, TYPE_CHECKING
 
 from .argvals import Argvals, DenseArgvals, IrregularArgvals
 from .values import Values, DenseValues, IrregularValues
@@ -43,6 +35,9 @@ from ..misc.utils import _outer
 
 if TYPE_CHECKING:
     from .basis import Basis
+
+
+NDArrayFloat = npt.NDArray[np.float64]
 
 
 ###############################################################################
@@ -78,14 +73,14 @@ def _tensor_product(
 
 
 def _smooth_covariance(
-    covariance_matrix: npt.NDArray[np.float64],
+    covariance_matrix: NDArrayFloat,
     argvals: DenseArgvals,
     points: DenseArgvals,
     method_smoothing: str = "LP",
     remove_diagonal: bool = True,
-    weights: npt.NDArray[np.float64] | None = None,
-    **kwargs,
-):
+    weights: NDArrayFloat | None = None,
+    **kwargs: Any,
+) -> NDArrayFloat:
     """Smooth the covariance.
 
     Parameters
@@ -176,11 +171,11 @@ def _smooth_covariance(
 
 
 def _estimate_noise_variance_with_covariance(
-    raw_diagonal: npt.NDArray[np.float64],
-    smooth_diagonal: npt.NDArray[np.float64],
+    raw_diagonal: NDArrayFloat,
+    smooth_diagonal: NDArrayFloat,
     argvals: DenseArgvals,
     points: DenseArgvals,
-):
+) -> float:
     """Estimate the variance of the noise using the covariance diagonal.
 
     Parameters
@@ -223,7 +218,7 @@ def _estimate_noise_variance_with_covariance(
         method="trapz",
     )
 
-    return np.maximum(2 * temp / points.range()["input_dim_0"], 0)
+    return np.maximum(2 * temp / points.range()["input_dim_0"], 0)  # type:ignore
 
 
 ###############################################################################
@@ -237,7 +232,7 @@ class FunctionalData(ABC):
         Number of observations of the functional data.
     n_dimension: int
         Number of input dimension of the functional data.
-    n_points: Tuple[int, ...] | Dict[int, Tuple[int, ...]]
+    n_points: tuple[int, ...] | dict[int, tuple[int, ...]]
         Number of sampling points.
 
     """
@@ -245,7 +240,7 @@ class FunctionalData(ABC):
     ###########################################################################
     # Checkers
     @staticmethod
-    def _check_same_type(*fdata: Type[FunctionalData]) -> None:
+    def _check_same_type(*fdata: FunctionalData) -> None:
         """Raise an error if elements in `fdata` have different type.
 
         Raises
@@ -259,7 +254,7 @@ class FunctionalData(ABC):
             raise TypeError("Elements do not have the same types.")
 
     @staticmethod
-    def _check_same_nobs(*fdata: Type[FunctionalData]) -> None:
+    def _check_same_nobs(*fdata: FunctionalData) -> None:
         """Raise an arror if elements in `fdata` have different number of obs.
 
         Raises
@@ -273,7 +268,7 @@ class FunctionalData(ABC):
             raise ValueError("Elements do not have the same number of observations.")
 
     @staticmethod
-    def _check_same_ndim(*fdata: Type[FunctionalData]) -> None:
+    def _check_same_ndim(*fdata: FunctionalData) -> None:
[…]
