+++ FDApy/preprocessing/dim_reduction/fcp_tpa.py
@@ -21,13 +21,15 @@ from ...representation.functional_data import DenseFunctionalData
 from ...misc.utils import _eigh
 
 
+NDArrayFloat = npt.NDArray[np.float64]
+
 ##############################################################################
 # Utility functions
 
 
 def _initialize_vectors(
     shape: Tuple[int, int, int]
-) -> Tuple[npt.NDArray, npt.NDArray, npt.NDArray]:
+) -> Tuple[NDArrayFloat, NDArrayFloat, NDArrayFloat]:
     r"""Init u, v and w in the FCP-TPA algorithm.
 
     Parameters
@@ -37,7 +39,7 @@ def _initialize_vectors(
 
     Returns
     -------
-    Tuple[npt.NDArray, npt.NDArray, npt.NDArray]
+    Tuple[NDArrayFloat, NDArrayFloat, NDArrayFloat]
         A tuple containing normed vectors for the initialization of the FCP-TPA
         algorithm.
 
@@ -48,7 +50,7 @@ def _initialize_vectors(
 
 def _initalize_output(
     shape: Tuple[int, int, int], n_components: int
-) -> Tuple[npt.NDArray, List[npt.NDArray]]:
+) -> Tuple[NDArrayFloat, List[NDArrayFloat]]:
     """Init coefficients and u, v and w eigenvectors matrices.
 
     Parameters
@@ -60,7 +62,7 @@ def _initalize_output(
 
     Returns
     -------
-    Tuple[npt.NDArray, List[npt.NDArray]]
+    Tuple[NDArrayFloat, List[NDArrayFloat]]
         A tuple containing initialized matrices for the results of the FCP-TPA
         algorithm.
 
@@ -71,8 +73,8 @@ def _initalize_output(
 
 
 def _eigendecomposition_penalty_matrices(
-    penalty_matrices: Dict[str, npt.NDArray[np.float64]]
-) -> Dict[str, Tuple[npt.NDArray, npt.NDArray]]:
+    penalty_matrices: Dict[str, NDArrayFloat]
+) -> Dict[str, Tuple[NDArrayFloat, NDArrayFloat]]:
     """Compute eigendecomposition of penalty matrices in the FCP-TPA algorithm.
 
     Parameters
@@ -84,7 +86,7 @@ def _eigendecomposition_penalty_matrices(
 
     Returns
     -------
-    Dict[str, Tuple[npt.NDArray, npt.NDArray]]
+    Dict[str, Tuple[NDArrayFloat, NDArrayFloat]]
         A dictionary where each entry contains the eigenvalues and eigenvectors
         of the penalty matrix for each dimension of the image. The eigenvalues
         are sorted in descending order.
@@ -96,9 +98,9 @@ def _eigendecomposition_penalty_matrices(
 def _gcv(
     alpha: float,
     dimension_length: int,
-    vector: npt.NDArray[np.float64],
+    vector: NDArrayFloat,
     smoother: float,
-    rayleigh: npt.NDArray[np.float64],
+    rayleigh: NDArrayFloat,
 ) -> float:
     r"""Generalized cross-validation for the FCP-TPA algortihm.
 
@@ -155,12 +157,12 @@ def _gcv(
 
 def _find_optimal_alpha(
     alpha_range: Tuple[float, float],
-    data: npt.NDArray[np.float64],
-    u: npt.NDArray[np.float64],
-    v: npt.NDArray[np.float64],
+    data: NDArrayFloat,
+    u: NDArrayFloat,
+    v: NDArrayFloat,
     alpha: float,
-    penalty_matrix: npt.NDArray[np.float64],
-    eigencomponents: Tuple[npt.NDArray[np.float64], npt.NDArray[np.float64]],
+    penalty_matrix: NDArrayFloat,
[…]
