### 2025-02-04 411da26
Re-order examples in documentation
