+++ examples/representation/plot_irregular_functional.py
@@ -2,7 +2,6 @@
 Representation of univariate and irregular functional data
 ==========================================================
 
-Examples of representation of univariate and irregular functional data.
 """
 
 # Author: Steven Golovkine <steven_golovkine@icloud.com>
@@ -16,26 +15,13 @@ from FDApy.representation import DenseArgvals, IrregularArgvals, IrregularValues
 from FDApy.visualization import plot
 
 ###############################################################################
-# The representation of irregular functional data
-# -----------------------------------------------
-# We are showing the building blocks of the representation of univariate and
-# irregular functional data. To define a ``FunctionalData`` object, we need a
-# set of ``argvals`` (the sampling points of the curves) and a set of
-# ``values`` (the observed points of the curves). The sampling points of the
-# data are defined as a dictionary where each entry is another dictionary that
-# represents an input dimension (one entry corresponds to curves, two entries
-# correspond to surface, ...). Each entry of the dictionary corresponds to the
-# sampling points of one observation.
-# The values of the functional data are defined in a dictionary where each
-# entry represents an observation as an ``npt.NDArray``. Each entry should have
-# the same dimension has the corresponding entry in the ``argvals`` dictionary.
+# In this section, we are showing the building blocks of the representation of univariate and irregular functional data. To define a :class:`~FDApy.representation.IrregularFunctionalData` object, we need a set of :class:`~FDApy.representation.IrregularArgvals` (the sampling points of the curves) and a set of :class:`~FDApy.representation.IrregularValues` (the observed points of the curves). The sampling points of the data are defined as a dictionary where each entry corresponds to an observation. Each entry of the dictionary corresponds to the sampling points of one observation and is represented as a :class:`~FDApy.representation.DenseArgvals`. The values of the functional data are defined in a dictionary where each entry represents an observation as an :class:`~numpy.ndarray`. Each entry should have the same dimension has the corresponding entry in the :class:`~FDApy.representation.IrregularArgvals` dictionary.
 
 ###############################################################################
 # For unidimensional functional data
 # ----------------------------------
-# First, we will define unidimensional functional data. We consider twenty
-# sampling points for the first observations and fifteen sampling points for
-# the second observations.
+# First, we will define unidimensional irregular functional data. We represent two observations of a functional data irregularly sampled on a set of points. The first observation in sampled on :math:`20` points between :math:`0` and :math:`1` and the second observation is sampled on :math:`15` points between :math:`0.2` and :math:`0.8`. The values of the functional data is a dictionary where each entry corresponds to the observed values of one observation.
+
 argvals = IrregularArgvals(
     {
         0: DenseArgvals({"input_dim_0": np.linspace(0, 1, num=20)}),
@@ -57,9 +43,8 @@ _ = plot(fdata)
 ###############################################################################
 # For two-dimensional functional data
 # -----------------------------------
-# Second, we will defined two-dimensional functional data. We consider a grid
-# of :math:`20 \times 20` sampling points for the first observation and a grid
-# of :math:`15 \times 15` sampling points for the second observation.
+# Second, we will consider two-dimensional functional data where the observations are not sampled on the same grid. We represent two observations of a functional data irregularly sampled on a set of points. The first observation is sampled on a grid of :math:`20 \times 20` sampling points and the second observation is sampled on a grid of :math:`15 \times 15` sampling points. The values of the functional data is a dictionary where each entry corresponds to the observed values of one observation.
+
 argvals = IrregularArgvals(
     {
         0: DenseArgvals(
@@ -90,3 +75,8 @@ X = IrregularValues(
 fdata = IrregularFunctionalData(argvals=argvals, values=X)
 
 _ = plot(fdata)
+
+###############################################################################
+# For higher-dimensional functional data
[…]
