### 2025-02-04 93ab9e8
Examples representation (#25)
