+++ examples/misc/README.rst
@@ -1,0 +1,7 @@
+.. _misc_examples:
+
+Miscellaneous
+-------------
+
+These examples illustrate diverse functionalities of the package that are not directly related to the main modules and to functional data analysis. They could be useful for the user to understand how to use the package in a more general context.
+
