### 2025-02-05 4040e01
Examples misc (#25)
