+++ examples/misc/plot_psplines_1d.py
@@ -2,11 +2,7 @@
 Smoothing of 1D data using P-Splines
 ====================================
 
-Examples of smoothing of one-dimensional data using P-Splines.
 """
-
-###############################################################################
-#
 
 # Author: Steven Golovkine <steven_golovkine@icloud.com>
 # License: MIT
@@ -15,7 +11,15 @@ Examples of smoothing of one-dimensional data using P-Splines.
 import matplotlib.pyplot as plt
 import numpy as np
 
-from FDApy.preprocessing.smoothing.psplines import PSplines
+from FDApy.preprocessing import PSplines
+
+###############################################################################
+# The package includes a class to perform P-Splines smoothing. The class :class:`~FDApy.preprocessing.PSplines` allows to fit a P-Splines regression to a functional data object. P-Splines regression is a non-parametric method that fits a spline to the data. The spline is defined by a basis of B-Splines. The B-Splines basis is defined by a set of knots. The P-Splines regression is a penalized regression that adds a discrete constraint to the fit. The influence of the penalty is controlled by the parameter `penalty`.
+#
+
+###############################################################################
+# We will show how to use the class :class:`~FDApy.preprocessing.PSplines` to smooth a one-dimensional dataset. We will simulate a dataset from a cosine function and add some noise. The goal is to recover the cosine function by fitting a P-Splines regression. The :class:`~FDApy.preprocessing.PSplines` class requires the specification of the number of segments and the degree of the B-Splines basis. The number of segments is used to define the number of knots. The degree of the B-Splines basis is used to define the order of the B-Splines basis. If the degree is set to :math:`0`, the B-Splines basis is a set of step functions. If the degree is set to :math:`1`, the B-Splines basis is a set of piecewise linear functions. If the degree is set to :math:`2`, the B-Splines basis is a set of piecewise quadratic functions. To fit the model, the method :meth:`~FDApy.preprocessing.PSplines.fit` requires the data and the penalty.
+#
 
 # Set general parameters
 rng = 42
@@ -28,7 +32,7 @@ y = np.cos(x) + 0.2 * rnorm(n_points)
 x_new = np.linspace(-2, 2, 51)
 
 ###############################################################################
-# Assess the influence of the degree of the B-Splines basis.
+# Here, we are interested in the influence of the degree of the B-Splines basis on the P-Splines regression. We will fit a P-Splines regression with degree :math:`0`, :math:`1` and :math:`2`. The number of segments is set to :math:`20` and the penalty is set to :math:`5`. We remark that the P-Splines regression with degree :math:`0` is not a good fit to data, while the P-Splines regression with degree :math:`1` or :math:`2` are roughly similar and recover the cosine function.
 #
 
 # Fit P-Splines regression with degree 0
@@ -46,7 +50,7 @@ ps = PSplines(n_segments=20, degree=2)
 ps.fit(y, x, penalty=5)
 y_pred_2 = ps.predict(x_new)
 
-###############################################################################
+# Plot results
 plt.scatter(x, y, c="grey", alpha=0.2)
 plt.plot(np.sort(x), np.cos(np.sort(x)), c="k", label="True")
 plt.plot(x_new, y_pred_0, c="r", label="Degree 0")
@@ -56,7 +60,8 @@ plt.legend()
 plt.show()
 
 ###############################################################################
-# Assess the influence of the penalty :math:`\lambda`.
[…]
