+++ examples/misc/plot_local_polynomials_1d.py
@@ -2,12 +2,7 @@
 Smoothing of 1D data using local polynomial regression
 ======================================================
 
-Examples of smoothing of one-dimensional data using local polynomial
-regression.
 """
-
-###############################################################################
-#
 
 # Author: Steven Golovkine <steven_golovkine@icloud.com>
 # License: MIT
@@ -17,6 +12,12 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from FDApy.preprocessing import LocalPolynomial
+
+###############################################################################
+# The package includes a class to perform local polynomial regression. The class :class:`~FDApy.preprocessing.LocalPolynomial` allows to fit a local polynomial regression to a functional data object. Local polynomial regression is a non-parametric method that fits a polynomial to the data in a local neighborhood of each point.
+
+###############################################################################
+# We will show how to use the class :class:`~FDApy.preprocessing.LocalPolynomial` to smooth a one-dimensional dataset. We will simulate a dataset from a cosine function and add some noise. The goal is to recover the cosine function by fitting a local polynomial regression. The :class:`~FDApy.preprocessing.LocalPolynomial` class requires the specification of the kernel, the bandwidth and the degree of the polynomial. The kernel is used to define the weights of the local regression. Four kernels are implemented: `gaussian`, `epanechnikov`, `tricube` and `bisquare`. The bandwidth is used to define the size of the local neighborhood. The degree of the polynomial is used to define the order of the polynomial to fit. If the degree is set to :math:`0`, the local regression is a local constant regression. If the degree is set to :math:`1`, the local regression is a local linear regression. If the degree is set to :math:`2`, the local regression is a local quadratic regression.
 
 # Set general parameters
 rng = 42
@@ -29,7 +30,7 @@ y = np.cos(x) + 0.2 * rnorm(n_points)
 x_new = np.linspace(-1, 1, 51)
 
 ###############################################################################
-# Assess the influence of the degree of the polynomials.
+# Here, we are interested in the influence of the degree of the polynomial on the local polynomial regression. We will fit a local polynomial regression with degree :math:`0`, :math:`1` and :math:`2`. The bandwidth is set to :math:`0.5` and the kernel is set to `epanechnikov`. We remark that the local polynomial regression with degree :math:`2` overfits the data, while the local polynomial regression with degree :math:`0` or :math:`1` roughly recover the cosine function.
 
 # Fit local polynomial regression with degree 0
 lp = LocalPolynomial(kernel_name="epanechnikov", bandwidth=0.5, degree=0)
@@ -43,8 +44,7 @@ y_pred_1 = lp.predict(y=y, x=x, x_new=x_new)
 lp = LocalPolynomial(kernel_name="epanechnikov", bandwidth=0.5, degree=2)
 y_pred_2 = lp.predict(y=y, x=x, x_new=x_new)
 
-
-###############################################################################
+# Plot results
 plt.scatter(x, y, c="grey", alpha=0.2)
 plt.plot(np.sort(x), np.cos(np.sort(x)), c="k", label="True")
 plt.plot(x_new, y_pred_0, c="r", label="Degree 0")
@@ -55,7 +55,7 @@ plt.show()
 
 
 ###############################################################################
-# Assess the influence of the bandwith :math:`\lambda`.
[…]
