### 2025-02-05 de9330f
Examples smoothing (#25)
