+++ examples/simulation/plot_simulation.py
@@ -2,7 +2,6 @@
 Simulation of functional data
 =============================
 
-Examples of simulation of functional data and the effect of adding noise and sparsification.
 """
 
 # Author: Steven Golovkine <steven_golovkine@icloud.com>
@@ -15,6 +14,12 @@ from FDApy.representation import DenseArgvals
 from FDApy.simulation import KarhunenLoeve
 from FDApy.visualization import plot
 
+###############################################################################
+# Two main features of functional data are noise and sparsity. The package provides 
+# methods to simulate noisy and sparse functional data. In this example, we will 
+# simulate noisy and sparse functional data using the Karhunen-Loève decomposition.
+
+
 # Set general parameters
 rng = 42
 n_obs = 10
@@ -23,6 +28,7 @@ n_obs = 10
 name = "bsplines"
 n_functions = 5
 argvals = DenseArgvals({"input_dim_0": np.linspace(0, 1, 101)})
+
 
 ###############################################################################
 # For one dimensional data
@@ -60,7 +66,7 @@ _ = plot(kl.noisy_data)
 # randomly removing a certain percentage of the sampling points. The percentage
 # of retain samplings points can be supplied by the user. In this example, the
 # retained number of observations will be different for each curve and be
-# randomly drawn between :math:`0.45` and :math:`0.55`.
+# randomly drawn between :math:`0.45` and :math:`0.55` (percentage :math:`\pm` epsilon).
 
 # Sparsify the data
 kl.sparsify(percentage=0.5, epsilon=0.05)
@@ -71,6 +77,11 @@ _ = plot(kl.sparse_data)
 ###############################################################################
 # For two dimensional data
 # ------------------------
+# We simulate :math:`N = 1` image on the two-dimensional observation grid
+# :math:`\{0, 0.01, 0.02, \cdots, 1\} \times \{0, 0.01, 0.02, \cdots, 1\}`,
+# based on the tensor product of the first :math:`K = 25` B-splines
+# basis functions on :math:`[0, 1] \times [0, 1]` and the variance of
+# the scores random variables equal to :math:`1`.
 
 # Parameters of the basis
 name = ("bsplines", "bsplines")
@@ -78,12 +89,6 @@ n_functions = (5, 5)
 argvals = DenseArgvals(
     {"input_dim_0": np.linspace(0, 1, 101), "input_dim_1": np.linspace(0, 1, 101)}
 )
-
-# We simulate :math:`N = 1` image on the two-dimensional observation grid
-# :math:`\{0, 0.01, 0.02, \cdots, 1\} \times \{0, 0.01, 0.02, \cdots, 1\}`,
-# based on the tensor product of the first :math:`K = 25` B-splines
-# basis functions on :math:`[0, 1] \times [0, 1]` and the variance of
-# the scores random variables equal to :math:`1`.
 
 kl = KarhunenLoeve(
     basis_name=name, n_functions=n_functions, argvals=argvals, random_state=rng
