+++ examples/simulation/plot_cluster.py
@@ -2,7 +2,6 @@
 Simulation of clusters of univariate functional data
 ====================================================
 
-Examples of simulation of clusters of univariate functional data.
 """
 
 # Author: Steven Golovkine <steven_golovkine@icloud.com>
@@ -14,6 +13,13 @@ import numpy as np
 from FDApy.representation import DenseArgvals
 from FDApy.simulation import KarhunenLoeve
 from FDApy.visualization import plot
+
+###############################################################################
+# The package provides a class to simulate clusters of univariate functional data based on the Karhunen-Loève decomposition. The class :class:`FDApy.simulation.KarhunenLoeve` allows to simulate functional data based on the truncated Karhunen-Loève representation of a functional process.
+
+###############################################################################
+# We simulate :math:`N = 20` curves on the one-dimensional observation grid :math:`\{0, 0.01, 0.02, \cdots, 1\}`, based on the first :math:`K = 25` Fourier basis functions on :math:`[0, 1]`. The clusters are defined through the coefficients in the Karhunen-Loève decomposition and parametrize using the `centers` parameter. The centers of the clusters are generated as Gaussian random variables with parameters defined by a `mean` and a `covariance`. We also consider an exponential decreasing of the eigenvalues.
+
 
 # Set general parameters
 rng = 42
@@ -33,14 +39,6 @@ mean = np.array([0, 0])
 covariance = np.array([[1, -0.6], [-0.6, 1]])
 centers = random_state.multivariate_normal(mean, covariance, size=n_functions)
 
-###############################################################################
-# We simulate :math:`N = 20` curves on the one-dimensional observation grid
-# :math:`\{0, 0.01, 0.02, \cdots, 1\}`, based on the first
-# :math:`K = 25` Fourier basis functions on :math:`[0, 1]`. The clusters are
-# defined through the coefficients in the Karhunen-Loève decomposition. The
-# centers of the clusters are generated as Gaussian random variables with
-# parameters defined by `mean` and `covariance`. We also consider an
-# exponential decreasing of the eigenvalues.
 kl = KarhunenLoeve(
     basis_name=name, argvals=argvals, n_functions=n_functions, random_state=rng
 )

[…]
