### 2025-02-06 b5e4d8d
Examples FPCA (#25)
