+++ examples/fpca/plot_mfpca_1d_sparse.py
@@ -2,12 +2,7 @@
 MFPCA of 1-dimensional sparse data
 ----------------------------------
 
-Example of multivariate functional principal components analysis of
-1-dimensional data.
 """
-
-###############################################################################
-#
 
 # Author: Steven Golovkine <steven_golovkine@icloud.com>
 # License: MIT
@@ -20,6 +15,9 @@ from FDApy.representation import DenseArgvals
 from FDApy.simulation import KarhunenLoeve
 from FDApy.preprocessing import MFPCA
 from FDApy.visualization import plot, plot_multivariate
+
+###############################################################################
+# In this section, we are showing how to perform a multivariate functional principal component analysis on one-dimensional sparse data using the :class:`~FDApy.preprocessing.MFPCA` class. We will compare two methods to perform the dimension reduction: the decomposition of the covariance operator and the decomposition of the inner-product matrix. We will use the first :math:`K = 3` principal components to reconstruct the curves.
 
 # Set general parameters
 rng = 42
@@ -37,15 +35,8 @@ argvals = [
 ]
 
 ###############################################################################
-# We simulate :math:`N = 50` curves of a 2-dimensional process. The first
-# component of the process is defined on the one-dimensional observation grid
-# :math:`\{0, 0.01, 0.02, \cdots, 1\}`, based on the first :math:`K = 5`
-# B-splines basis functions on :math:`[0, 1]` and the variance of the scores
-# random variables equal to :math:`1`. The second component of the process is
-# defined on the one-dimensional observation grid
-# :math:`\{0, 0.01, 0.02, \cdots, 1\}`, based on the first :math:`K = 5`
-# Fourier basis functions on :math:`[0, 1]` and the variance of the scores
-# random variables equal to :math:`1`.
+# We simulate :math:`N = 50` curves of a 2-dimensional process. The first component of the process is defined on the one-dimensional observation grid :math:`\{0, 0.01, 0.02, \cdots, 1\}`, based on the first :math:`K = 5` B-splines basis functions on :math:`[0, 1]` and the variance of the scores random variables equal to :math:`1`. The second component of the process is defined on the one-dimensional observation grid :math:`\{0, 0.01, 0.02, \cdots, 1\}`, based on the first :math:`K = 5` Fourier basis functions on :math:`[0, 1]` and the variance of the scores random variables equal to :math:`1`.
+
 kl = KarhunenLoeve(
     basis_name=name, n_functions=n_functions, argvals=argvals, random_state=rng
 )
@@ -55,13 +46,13 @@ data = kl.sparse_data
 
 _ = plot_multivariate(data)
 
-##############################################################################
-# Covariance decomposition
-# ------------------------
 
-# Perform multivariate FPCA with an estimation of the number of components by
-# the percentage of variance explained using a decomposition of the covariance
-# operator.
+###############################################################################
+# Estimation of the eigencomponents
+# ---------------------------------
+#
+# The :class:`~FDApy.preprocessing.MFPCA` class requires two parameters: the number of components to estimate and the method to use. The method parameter can be either `covariance` or `inner-product`. The first method estimates the eigenfunctions by decomposing the covariance operator, while the second method estimates the eigenfunctions by decomposing the inner-product matrix. In the case of a decomposition of the covariance operator, the method also requires the univariate expansions to estimate the eigenfunctions of each component. Here, we use the univariate functional principal component analysis with penalized splines to estimate the eigenfunctions of each component.
+
 univariate_expansions = [
[…]
