+++ examples/fpca/plot_fpca_1d.py
@@ -2,11 +2,7 @@
 FPCA of 1-dimensional data
 --------------------------
 
-Example of functional principal components analysis of 1-dimensional data.
 """
-
-###############################################################################
-#
 
 # Author: Steven Golovkine <steven_golovkine@icloud.com>
 # License: MIT
@@ -20,6 +16,10 @@ from FDApy.simulation import KarhunenLoeve
 from FDApy.preprocessing import UFPCA
 from FDApy.visualization import plot
 
+###############################################################################
+# In this section, we are showing how to perform a functional principal component on one-dimensional data using the :class:`~FDApy.preprocessing.UFPCA` class. We will compare two methods to perform the dimension reduction: the decomposition of the covariance operator and the decomposition of the inner-product matrix. We will use the first :math:`K = 5` principal components to reconstruct the curves.
+
+
 # Set general parameters
 rng = 42
 n_obs = 50
@@ -29,11 +29,10 @@ name = "fourier"
 n_functions = 25
 argvals = DenseArgvals({"input_dim_0": np.linspace(0, 1, 101)})
 
+
 ###############################################################################
-# We simulate :math:`N = 50` curves on the one-dimensional observation grid
-# :math:`\{0, 0.01, 0.02, \cdots, 1\}`, based on the first :math:`K = 25`
-# Fourier basis functions on :math:`[0, 1]` and the variance of the scores
-# random variables decreasing exponentially.
+# We simulate :math:`N = 50` curves on the one-dimensional observation grid :math:`\{0, 0.01, 0.02, \cdots, 1\}`, based on the first :math:`K = 25` Fourier basis functions on :math:`[0, 1]` and the variance of the scores random variables decreasing exponentially.
+
 kl = KarhunenLoeve(
     n_functions=n_functions, basis_name=name, argvals=argvals, random_state=rng
 )
@@ -42,73 +41,61 @@ kl.add_noise(noise_variance=0.05)
 data = kl.noisy_data
 
 _ = plot(data)
+plt.show()
 
 ###############################################################################
-# Covariance decomposition
-# ------------------------
+# Estimation of the eigencomponents
+# ---------------------------------
 #
-# We perform a univariate FPCA with a predefined number of components using a
-# decomposition of the covariance operator.
+# The :class:`~FDApy.preprocessing.UFPCA` class requires two parameters: the number of components to estimate and the method to use. The method parameter can be either `covariance` or `inner-product`. The first method estimates the eigenfunctions by decomposing the covariance operator, while the second method estimates the eigenfunctions by decomposing the inner-product matrix.
+
+# First, we perform a univariate FPCA using a decomposition of the covariance operator.
 ufpca_cov = UFPCA(n_components=5, method="covariance")
 ufpca_cov.fit(data)
 
-# Plot the eigenfunctions
+# Plot the eigenfunctions using the decomposition of the covariance operator.
 _ = plot(ufpca_cov.eigenfunctions)
-
+plt.show()
 
 ###############################################################################
-# We estimate the scores, which is the projection of the curves onto the
-# eigenfunctions, by numerical integration and using PACE.
+#
+
+# Second, we perform a univariate FPCA using a decomposition of the inner-product matrix.
+ufpca_innpro = UFPCA(n_components=5, method="inner-product")
+ufpca_innpro.fit(data)
+
+# Plot the eigenfunctions using the decomposition of the inner-product matrix.
+_ = plot(ufpca_innpro.eigenfunctions)
+plt.show()
+
+###############################################################################
+# Estimation of the scores
+# ------------------------
+#
[…]
