+++ examples/data_analysis/plot_cd4.py
@@ -1,6 +1,6 @@
 """
 CD4 dataset
------------
+============
 
 """
 
