+++ examples/fpca/plot_mfpca_1d_sparse.py
@@ -1,6 +1,6 @@
 """
 MFPCA of 1-dimensional sparse data
-----------------------------------
+==================================
 
 """
 
