+++ examples/data_analysis/plot_canadian_weather.py
@@ -1,6 +1,6 @@
 """
 Canadian weather dataset
-------------------------
+=========================
 
 """
 
