+++ examples/fpca/plot_fpca_1d.py
@@ -1,6 +1,6 @@
 """
 FPCA of 1-dimensional data
---------------------------
+===========================
 
 """
 
