+++ examples/fpca/plot_fpca_1d_sparse.py
@@ -1,6 +1,6 @@
 """
 FPCA of 1-dimensional sparse data
----------------------------------
+==================================
 
 """
 
