+++ examples/fpca/plot_fpca_2d.py
@@ -1,6 +1,6 @@
 """
 FPCA of 2-dimensional data
---------------------------
+===========================
 
 """
 
