### 2025-02-06 StevenGolovkine/FDApy#28 @StevenGolovkine
Hi @quantgirluk,

Thanks for your comments. I've changed the examples and try to provide more explanations (also in reference to #25).
Let me know if you require something else.
