### 2025-02-06 StevenGolovkine/FDApy#23 @StevenGolovkine
Hi @vnmabus,

Thanks for your comments. I was not really aware of this.
I dropped the support of Python 3.9 for Python 3.10 as a minimal requirement. I will work on the typing for the next release of the package. However, I'll need some time to do it.
Do you consider this as a requirement for the publication of the paper?
