### 2025-02-06 StevenGolovkine/FDApy#23 @vnmabus
> Do you consider this as a requirement for the publication of the paper?

Of course not. Type hints are optional in Python, after all. These are just feedback hints to improve that aspect of the package, as you seemed to want to properly type it. They are in no way mandatory.
