### 2025-02-06 StevenGolovkine/FDApy#26 @vnmabus
About 2, I am not sure how a long representation of irregular data implies a dependency on Pandas. In scikit-fda we achieve a similar representation with three NumPy arrays (function indices, input points, and values).

About 3, some vectorized computations that I think are useful for this kind of data are, for example, normal operations (e. g. adding/multiplying each function by a number, or by another irregular function of the same shape) or per-function reductions (e. g. summing the values of the functions, for example in order to compute a quadrature). With the current implementation it is always necessary to have at least one Python loop iterating over the functions, while with a "long format" in three NumPy arrays it is possible to use vectorized routines from NumPy to improve the performance of some of these operations.

Again, these are substantial changes for you to consider for the future, and won't affect JOSS publication from my part.
