### 2025-02-06 StevenGolovkine/FDApy#22 @vnmabus
> The formula of the norm is written in the doc.

Not for all representations, nor the base classes. I think I looked at one of these.

> I would say it is not really useful for univariate functional data. However, for multivariate functional data, it would make sense to standardize each individual feature if they have very different scale, even to run MFPCA, so the resulting functional eigencomponents won't represent only the features with the most variance.

I understand the reasoning. However, by using the functional mean and std, aren't you altering the functional nature of the data? As far as I know, in most functional datasets all values of the same function are in the same units (e. g. temperature in ºC for the Canadian Weather dataset). By standardizing using the mean function and mean std, each point of the function is now in a different scale. Shouldn't it be more logical to standardize using some "mean-like constant" and "std-like constant" in order to preserve the scale? I have not seen literature on this topic, which is why we didn't implement a similar feature in scikit-fda for now, but maybe you know this topic better.
