### 2025-02-07 StevenGolovkine/FDApy#26 @StevenGolovkine
I didn't know about skfda implementation. I think both work, as for some methods, you will need to split the arrays and have something similar than my representation (and for some methods, I need the long format and so a representation similar to yours). I wanted similar representation for DenseFunctionalData and IrregularFunctionalData, with two arguments: Argvals and Values.
However, you may need to split your data fewer times than the number of times I need to convert it to long format. In any case, thanks a lot for this, it makes me consider different perspectives on my implementation (which hasn't really been thought of by others)
