+++ FDApy/representation/functional_data.py
@@ -3258,7 +3258,13 @@ class BasisFunctionalData(FunctionalData):
         method_integration: str = "trapz",
         use_argvals_stand: bool = False,
     ) -> npt.NDArray[np.float64]:
-        """Norm of each observation of the data.
+        r"""Norm of each observation of the data.
+
+        For each observation in the data, it computes its norm defined
+        as
+
+        .. math::
+            \| X \| = \left\{\int_{\mathcal{T}} X(t)^2dt\right\}^{\frac12}.
 
         Parameters
         ----------
