### 2025-02-07 StevenGolovkine/FDApy#22 @StevenGolovkine
That's a good question. I would tend to agree with you. However as I was working on MFPCA, I found some papers about standardization using the functional std. So, I implemented it.
You can look [here](https://doi.org/10.51387/23-NEJSDS31) for the centering and [here](https://www.jstor.org/stable/24310959) for the standardization.
