### 2025-02-07 StevenGolovkine/FDApy#22 @vnmabus
Thanks for the explanation, and the linked papers. I think that the API reference is in a way better shape now, so you may close this issue if you wish.
