### 2025-02-14 StevenGolovkine/FDApy#25 @vnmabus
> It is not possible to estimate the scores for new data using the inner-product decomposition (only using the projection of the curves onto the eigenfunctions). As we already have the data at end, it is useless to pass it to the transform function. See my paper for more information: https://arxiv.org/abs/2306.12949

I guess that what you mean is that with the inner-product method the estimator is [transductive](https://scikit-learn.org/stable/glossary.html#term-transductive) and not [inductive](https://scikit-learn.org/stable/glossary.html#term-inductive). In that case, what scikit-learn does is to provide just a [`fit_transform`](https://scikit-learn.org/stable/glossary.html#term-fit_transform) method, and not a [`transform`](https://scikit-learn.org/stable/glossary.html#term-transform) method. I know you said that scikit-learn compatibility is out-of-scope for now, but I would advise you to keep at least the same semantics if you are going to use the same nomenclature. It would minimize user's confusion, and will require less changes if you add that compatibility in the future.

> I created a new section with methods not directly related to functional data (Local Polynomial and P-Splines) and moved the smoothing of functional data into the Smoothing section.

In which sense are they not directly related with functional data? Surely for applying *any* smoothing techniques you require that your data are continuous functions.
