### 2025-02-17 StevenGolovkine/FDApy#25 @vnmabus
Thanks for your answers @StevenGolovkine . You can close this issue if you want.

I will check the tests of the package when I have time (I forgot to do it the first time) and if everything works nicely the review would be finished from my part.  Thank you for your time making this software and for your answers!
