### 2025-02-19 42a6e35
Fix the tests
Fix the following error when running `pytest`:
```
TypeError: simpson() takes 1 positional argument but 2 were given
```

Fix also the following formatting errors when
running `flake8 FDApy tests`:

```
FDApy/representation/functional_data.py:439:1: W293 blank line contains whitespace
FDApy/representation/functional_data.py:734:1: W293 blank line contains whitespace
tests/test_psplines.py:179:1: W293 blank line contains whitespace
tests/test_psplines.py:229:1: W293 blank line contains whitespace
```
