### 2025-02-20 #7526 @mstimberg
@vnmabus Many thanks for your feedback (and for your constructive review, obviously)!

@quantgirluk How are things looking from your side – are you satisfied with the changes @StevenGolovkine made in response to your issues (StevenGolovkine/FDApy#27 and StevenGolovkine/FDApy#28)? Is there anything else in the software or the manuscript that you think could be improved?
