### 2025-02-25 d194feb
JOSS: normalize capitalization in references
