+++ joss/paper.bib
@@ -1,16 +1,16 @@
 @techreport{vanrossumPEP8StyleGuide2001,
   type = {{{PEP}}},
-  title = {{{PEP8}} - {{Style}} Guide for {{Python}} Code},
+  title = {PEP8 - Style Guide for {Python} Code},
   author = {{van Rossum}, Guido and Warsaw, Barry and Coghlan, Alyssa},
   year = {2001},
 }
 
 @book{ramsayFunctionalDataAnalysis2005,
-  title = {Functional {{Data Analysis}}},
+  title = {Functional Data Analysis},
   author = {Ramsay, J. O. and Silverman, B. W.},
   year = {2005},
-  series = {Springer {{Series}} in {{Statistics}}},
-  publisher = {{Springer}},
+  series = {Springer Series in Statistics},
+  publisher = {Springer},
   address = {{New York, NY}},
   urldate = {2023-07-07},
   langid = {english},
@@ -19,20 +19,20 @@
 }
 
 @book{horvathInferenceFunctionalData2012a,
-  title = {Inference for {{Functional Data}} with {{Applications}}},
+  title = {Inference for Functional Data with Applications},
   author = {Horv{\'a}th, Lajos and Kokoszka, Piotr},
   year = {2012},
-  series = {Springer {{Series}} in {{Statistics}}},
+  series = {Springer Series in Statistics},
   volume = {200},
-  publisher = {{Springer}},
-  address = {{New York, NY}},
+  publisher = {Springer},
+  address = {New York, NY},
   urldate = {2023-09-19},
   langid = {english},
   keywords = {functional-data-analysis,functional-time-series,hilbert-space-theory},
 }
 
 @book{kokoszkaIntroductionFunctionalData2017,
-  title = {Introduction to {{Functional Data Analysis}}},
+  title = {Introduction to Functional Data Analysis},
   author = {Kokoszka, Piotr and Reimherr, Matthew},
   year = {2017},
   month = sep,
@@ -52,7 +52,7 @@
 }
 
 @article{happMultivariateFunctionalPrincipal2018,
-  title = {Multivariate {{Functional Principal Component Analysis}} for {{Data Observed}} on {{Different}} ({{Dimensional}}) {{Domains}}},
+  title = {Multivariate Functional Principal Component Analysis for Data Observed on Different (Dimensional) Domains},
   author = {Happ, Clara and Greven, Sonja},
   year = {2018},
   month = apr,
@@ -84,20 +84,6 @@
   doi = {10.1109/camsap.2013.6714047}
 }
 
-@misc{ramos-carrenoScikitfdaPythonPackage2023,
-  title = {Scikit-Fda: {{A Python Package}} for {{Functional Data Analysis}}},
-  shorttitle = {Scikit-Fda},
-  author = {{Ramos-Carre{\~n}o}, Carlos and Torrecilla, Jos{\'e} Luis and {Carbajo-Berrocal}, Miguel and Marcos, Pablo and Su{\'a}rez, Alberto},
-  year = {2023},
-  month = jun,
-  number = {arXiv:2211.02566},
-  eprint = {2211.02566},
-  primaryclass = {cs, stat},
-  publisher = {arXiv},
-  doi = {10.48550/arXiv.2211.02566},
-  urldate = {2024-01-15},
-}
-
 @misc{loningSktimeSktimeV02022,
   title = {Sktime/Sktime: V0.13.4},
   shorttitle = {Sktime/Sktime},
@@ -110,7 +96,7 @@
 }
 
 @article{tavenardTslearnMachineLearning2020,
[…]
