### 2025-02-25 #7526 @mstimberg
@StevenGolovkine I made a small PR in StevenGolovkine/FDApy#36 with some corrections to the bibliography. Please merge if there's nothing you disagree with.
