+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog
 =========
 
+Version 1.0.3 (2025-02-25)
+==========================
+
+- JOSS paper
+
 Version 1.0.2 (2024-09-24)
 ==========================
 
