+++ CITATION.cff
@@ -1,0 +1,15 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+  - family-names: "Golovkine"
+    given-names: "Steven"
+    orcid: "https://orcid.org/0000-0002-5994-2671"
+    affiliation: "University of Limerick"
+    email: steven_golovkine@icloud.com
+title: "StevenGolovkine/FDApy: a Python package for functional data"
+date-released: 2025-02-28
+url: "https://github.com/StevenGolovkine/FDApy"
+license: MIT
+keywords:
+  - functional data analysis
+  - Python
