### 2025-02-28 2a05315
Update CITATION.cff
