+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+
+Version 1.0.4 (2025-02-28)
+==========================
+
+- Zenodo DOI
+
 Version 1.0.3 (2025-02-25)
 ==========================
 
