+++ CHANGELOG.rst
@@ -3,11 +3,6 @@ Changelog
 =========
 
 
-Version 1.0.4 (2025-02-28)
-==========================
-
-- Zenodo DOI
-
 Version 1.0.3 (2025-02-25)
 ==========================
 
