### 2025-02-28 #7526 @StevenGolovkine
Hi @mstimberg,

Author, affiliation and ORCID are correct. The version that should be used in the JOSS paper is **v1.0.3**.
The release has been archived on Zenodo (https://doi.org/10.5281/zenodo.14944699).
Title, author, ORCID in the archive match those in the JOSS paper.
The license is the same in the archive and the software (MIT).
