+++ joss/paper.bib
@@ -41,15 +41,6 @@
   keywords = {functional-data-analysis,functional-regression,functional-time-series,hilbert-space-theory},
 }
 
-@article{wangFaultDetectionBatch2015,
-  title = {Fault Detection of Batch Processes Based on Multivariate Functional Kernel Principal Component Analysis},
-  author = {Wang, Huangang and Yao, Ma},
-  year = {2015},
-  month = dec,
-  journal = {Chemometrics and Intelligent Laboratory Systems},
-  volume = {149},
-  pages = {78--89},
-}
 
 @article{happMultivariateFunctionalPrincipal2018,
   title = {Multivariate Functional Principal Component Analysis for Data Observed on Different (Dimensional) Domains},
