### 2025-02-28 #7526 @mstimberg
Great, thanks. One final minor detail – could you please change the name on the Zenodo archive so that it is exactly the same as the paper (i.e. "StevenGolovkine/FDApy: …" → "FDApy: …")? No need to do a new archive/release, you can update the metadata of the existing one.
