### 2025-02-28 #7526 @mstimberg
Hey @StevenGolovkine, it seems that the bib file still has several duplicated entries, could you please remove them?
