+++ joss/paper.bib
@@ -1,9 +1,3 @@
-@techreport{vanrossumPEP8StyleGuide2001,
-  type = {{{PEP}}},
-  title = {PEP8 - Style Guide for {Python} Code},
-  author = {{van Rossum}, Guido and Warsaw, Barry and Coghlan, Alyssa},
-  year = {2001},
-}
 
 @book{ramsayFunctionalDataAnalysis2005,
   title = {Functional Data Analysis},
@@ -18,28 +12,6 @@
   doi = {10.1007/b98888}
 }
 
-@book{horvathInferenceFunctionalData2012a,
-  title = {Inference for Functional Data with Applications},
-  author = {Horv{\'a}th, Lajos and Kokoszka, Piotr},
-  year = {2012},
-  series = {Springer Series in Statistics},
-  volume = {200},
-  publisher = {Springer},
-  address = {New York, NY},
-  urldate = {2023-09-19},
-  langid = {english},
-  keywords = {functional-data-analysis,functional-time-series,hilbert-space-theory},
-}
-
-@book{kokoszkaIntroductionFunctionalData2017,
-  title = {Introduction to Functional Data Analysis},
-  author = {Kokoszka, Piotr and Reimherr, Matthew},
-  year = {2017},
-  month = sep,
-  publisher = {{Chapman and Hall/CRC}},
-  address = {{New York}},
-  keywords = {functional-data-analysis,functional-regression,functional-time-series,hilbert-space-theory},
-}
 
 
 @article{happMultivariateFunctionalPrincipal2018,
@@ -75,16 +47,6 @@
   doi = {10.1109/camsap.2013.6714047}
 }
 
-@misc{loningSktimeSktimeV02022,
-  title = {Sktime/Sktime: V0.13.4},
-  shorttitle = {Sktime/Sktime},
-  author = {L{\"o}ning, Markus and Kir{\'a}ly, Franz and Bagnall, Tony and Middlehurst, Matthew and Ganesh, Sajaysurya and Oastler, George and Lines, Jason and Walter, Martin and ViktorKaz and Mentel, Lukasz and {chrisholder} and Tsaprounis, Leonidas and RNKuhns and Parker, Mirae and Owoseni, Taiwo and Rockenschaub, Patrick and {danbartl} and {jesellier} and {eenticott-shell} and Gilbert, Ciaran and Bulatova, Guzal and Lovkush and Sch{\"a}fer, Patrick and Khrapov, Stanislav and Buchhorn, Katie and Take, Kejsi and Subramanian, Shivansh and Meyer, Svea Marie and AidenRushbrooke and {rice}, Beth},
-  year = {2022},
-  month = sep,
-  doi = {10.5281/zenodo.7117735},
-  urldate = {2024-03-21},
-  howpublished = {Zenodo},
-}
 
 @article{tavenardTslearnMachineLearning2020,
   title = {Tslearn, A Machine Learning Toolkit for Time Series Data},
@@ -99,48 +61,6 @@
   url = {http://jmlr.org/papers/v21/20-091.html}
 }
 
-@misc{ramsayFdaFunctionalData2023,
-  title = {Fda: Functional Data Analysis},
-  shorttitle = {Fda},
-  author = {Ramsay, James and Hooker, Giles and Graves, Spencer},
-  year = {2023},
-  month = may,
-  urldate = {2024-01-10},
-}
-
-@book{ramsayFunctionalDataAnalysis2009,
-  title = {Functional Data Analysis with {{R}} and {{MATLAB}}},
-  author = {Ramsay, James and Hooker, Giles and Graves, Spencer},
-  year = {2009},
-  publisher = {Springer},
-  address = {New York, NY},
[…]
