### 2025-02-28 #7526 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
