### 2025-03-03 #7526 @crvernon
:wave: @StevenGolovkine - I just need you to address the following before I move to accept this for publication:

- [ ] Please remove the formatting from your package name in the title of your paper.  This will cause display issues when JOSS publishes.
- [ ] LINE 9:  You are missing a space between the two references:  `@ramsayFunctionalDataAnalysis2005,@horvathInferenceFunctionalData2012a,`

Let me know when these have been addressed and I'll move to accept.  Thanks.
